% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{add_intensities}
\alias{add_intensities}
\title{Adding intensities}
\usage{
add_intensities(df, adducts )
}
\arguments{
\item{df}{input dataframe}

\item{adducts}{data frame adducts}
}
\description{
A list of annotation information of isotopes
}
