% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addFeatureData}
\alias{addFeatureData}
\title{Add feature data to the experiment}
\usage{
addFeatureData(exp, data = NULL)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{data}{Dataframe with featurenames as rows}
}
\description{
Add feature data to the experiment
}
