% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{addIntensities}
\alias{addIntensities}
\title{Adding intensities}
\usage{
addIntensities(data, iso_df)
}
\arguments{
\item{data}{data.frame, containing mz and intensity information}

\item{iso_df}{data.frame that contains possible mono-isotopic ion and isotope based on solely mass differences}
}
\description{
Adds an intensity column for mono-isotopic ion and isotope ions
}
