% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addMetadata}
\alias{addMetadata}
\title{Add metadata to the SummarizedExperiment}
\usage{
addMetadata(exp, metaList)
}
\arguments{
\item{exp}{SummarizedExperiment object obtained after alignment}

\item{metaList}{Named list of metadata fields that can be of any type.}
}
\value{
SummarizedExperiment object with added metadata in the \link{metadata}
slot.
}
\description{
the \link{metadata} slot of the SummarizedExperiment can hold any
type data and is often used for describing the experiment or data types that
don't fit in the \link{assay}, \link{colData}, or \link{rowData} slots. This function aids
in setting metadata.
}
\details{
Metadata in a SummarizedExperiment is used for miscelaneous data
that describe the experiment. Examples are the datetime of when the
experiment was made or model information (which is what it is used for in
sumR). However, additional data can be set as well with this function.
}
\examples{
# Read data
data("sumRnegative")

# Set the current time in the experiment
sumRnegative <- addMetadata(sumRnegative, list(Time = lubridate::now()))

# Show the metadata slot
metadata(sumRnegative)
}
