% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addPeakData}
\alias{addPeakData}
\title{Add metadata of peaks to the SummarizedExperiment}
\usage{
addPeakData(exp, data = NULL)
}
\arguments{
\item{exp}{SummarizedExperiment object obtained after alignment}

\item{data}{Dataframe with peaknames as rows and attributes as columns.}
}
\value{
SummarizedExperiment object with added peak data in the \link{rowData}
slot of the SummarizedExperiment
}
\description{
Peaks are the rows in a SummarizedExperiment. This function
allows you to set metadata about these peaks. In most cases, this is handled
by sumR itself, however, any additional annotation data may be set using
this function.
}
\details{
Metadata of peaks are stored in the rowData of the
SummarizedExperiment. These may contain metrics of peaks like mz and rt or
other summarized information like RSD, or other ratios. Therefore, rowData
is often set by sumR itself. In some cases, like compound annotation,
identifiers or other metadata it may be set by the user. This function aids
in doing so.
}
\examples{
# Read data
data("sumRnegative")

# Setup data.frame of polarity
df <- data.frame(Polarity = rep("positive",
                 nrow(sumRnegative)),
                 row.names = rownames(sumRnegative))

# Adding polarity to the peaks
sumRnegative <- addPeakData(sumRnegative, df)

# Showing rowData
rowData(sumRnegative)
}
