% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{addSampleBlankRatio}
\alias{addSampleBlankRatio}
\title{Add the Sample/Blank ratio for each compound}
\usage{
addSampleBlankRatio(exp, assay = 1, typeColumn = "Type")
}
\arguments{
\item{exp}{SummarizedExperiment Object}

\item{typeColumn}{Name of the column in `colData(exp)` that represents
the sample type and contains at least the values `SAMPLE` and `BLANK`.}
}
\description{
Add the Sample/Blank ratio for each compound
}
