% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addSampleData}
\alias{addSampleData}
\title{Add sample data to the experiment}
\usage{
addSampleData(exp, data = NULL)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{data}{Dataframe with samplenames as rows}
}
\description{
Add sample data to the experiment
}
