% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{adductFinding}
\alias{adductFinding}
\title{Adduct finding}
\usage{
adductFinding(adduct_df, max_diff_mat, min_diff_mat, mz_vector)
}
\arguments{
\item{adduct_df}{data.frame containing adduct information, name and mass columns}

\item{max_diff_mat}{matrix with maximum difference between mz pairs}

\item{min_diff_mat}{matrix with minumum difference between mz pairs}

\item{mz_vector}{vector of mz values}
}
\description{
Searches for adducts based on mass differences
}
