% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{Alkyne_n}
\alias{Alkyne_n}
\title{Estimation of number of carbon in Alkyne}
\usage{
Alkyne_n(n)
}
\arguments{
\item{mz}{Mass of interest in m/z}
}
\description{
Calculates the number of carbons in C(n)H(2n-2)
}
