% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{annnotate_formula}
\alias{annnotate_formula}
\title{Annotation for formula}
\usage{
annnotate_formula(final_df,list_annotate)
}
\arguments{
\item{final_df}{final dataframe}

\item{list_annotate}{A list of annotation information of isotopes}
}
\description{
Adds column with chemical formula
}
