% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{annotate_add}
\alias{annotate_add}
\title{Annotation for adducts incl isotopes}
\usage{
annotate_add(mz_vector, ppm)
}
\arguments{
\item{mz_vector}{mz value vector}

\item{ppm}{error part per million}
}
\description{
A list of annotation information of adducts)
}
