% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{annotate_iso}
\alias{annotate_iso}
\title{Annotation for isotopes}
\usage{
annotate_iso(mz_vector, ppm , z )
}
\arguments{
\item{mz_vector}{mz value vector}

\item{ppm}{error part per million}

\item{z}{charge values}
}
\description{
A list of annotation information of isotopes
}
