% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{anova_tukey}
\alias{anova_tukey}
\title{anova-tukey test}
\usage{
anova_tukey(df, classifiers)
}
\arguments{
\item{df}{transposed dataframe with m/z as columns}

\item{classifiers}{a factor vector with the classes of the samples}
}
\description{
anova test followed by tukey HSD test as post-hoc
}
