% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{aovTukeyTest}
\alias{aovTukeyTest}
\title{anova-tukey test}
\usage{
aovTukeyTest(exp, assay = 1, classifiers = metadata(exp)$phenotype)
}
\description{
anova test followed by tukey HSD test as post-hoc
}
