% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autoScale}
\alias{autoScale}
\title{Auto-scale with optional log transform an assay}
\usage{
autoScale(exp, assay = 1, log = TRUE, base = exp(1), setDefault = TRUE)
}
\arguments{
\item{exp}{SummarizedExperiment object obtained after imputation}

\item{assay}{Name or index of the assay to use. Defaults to the assay at
index 1.}

\item{log}{Boolean value, should log transformation be applied before
scaling? Defaults to TRUE}

\item{base}{If \code{log} is set to TRUE, what base should be used for the log
transformation? Defaults to \code{exp(1)}}

\item{setDefault}{Boolean value, should the new assay be set as default
assay on index 1? Defaults to TRUE}
}
\description{
Auto-scale with optional log transform an assay
}
