% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{autoTest}
\alias{autoTest}
\title{Detect and execute an appropriate statistical test}
\usage{
autoTest(exp, assay = 1, classifiers = metadata(exp)$phenotype)
}
\arguments{
\item{exp}{SummarizedExperiment with a (auto)scaled assay.}

\item{assay}{Name or index of the assay with (auto)scaled values. Defaults
to the first assay (index 1).}

\item{classifiers}{Column in the \link{colData} that describes the phenotype or
predictor class used for statistical testing and modelling. Defaults to the
phenotype set using \link{setPhenotype}.}
}
\description{
Detect and execute an appropriate statistical test
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Detect and execute appropriate test
sumRnegative <- autoTest(sumRnegative)
}
