% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{barPlot}
\alias{barPlot}
\title{Bar plot visualization}
\usage{
barPlot(final_df)
}
\arguments{
\item{final_df}{a data.frame of the mono-isotopic and isotopic peaks pair's with id column}
}
\description{
Displays a bar plot for the different types of isotopic status'
}
