% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{BG_filter}
\alias{BG_filter}
\title{BG filter}
\usage{
BG_filter(
  dataframe,
  limit = 2.5,
  blank_regex = "blank",
  mz_regex = "mz",
  sample_regex = "^d|^ec|qc"
)
}
\arguments{
\item{mz_column}{}
}
\description{
it is used to conduct background removal).
It filters a dataframe of m/z values and intensities.
This function allows the user to remove m/z peaks of measured samples, 
if they correspond to the same m/z of a blank within a specified range 
(for example within 150% of the blank intensity).
}
\examples{

}
\author{
amar
}
