% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{bgFilter}
\alias{bgFilter}
\title{Blank filter to remove background}
\usage{
bgFilter(dataframe, limit = 2.5, blank_regex = "blank", sample_regex = "^d|^ec|qc")
}
\arguments{
\item{dataframe}{a data.frame of m/z values and intensities including blank samples}

\item{limit}{intensity ratio limit, for example within 150 percent of the blank intensity = 1.5}

\item{blank_regex}{regex of blank columns, for example "blank"}

\item{sample_regex}{regex of sample columns for example}
}
\description{
It is used to conduct background removal, it filters a data frame of m/z values and intensities
}
