% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{bigMerge}
\alias{bigMerge}
\title{Big Merging function}
\usage{
bigMerge(isotope_valid_df, data, copy_df)
}
\arguments{
\item{isotope_valid_df}{a data.frame of valid annotate information for isotopes}

\item{data}{the input data.frame, with min_error and max_error columns bind}

\item{copy_df}{a data.frame copy of the input data frame}
}
\description{
Merging all different isotope data frames into one
}
