% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{binCells}
\alias{binCells}
\title{Bin cells}
\usage{
binCells(
  spectraList,
  cellData = NULL,
  phenotype = NULL,
  tolerance = 0.002,
  filter = TRUE
)
}
\arguments{
\item{spectraList}{List of spectra}
}
\description{
Bin cells
}
