% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Setting of the bin boundaries}
\usage{
binning(mass, intensities, samples, tolerance = 5e-06)
}
\arguments{
\item{mass}{obtained from the input files}

\item{intensities}{obtained from the input files}

\item{samples}{obtained from the input files}

\item{tolerance}{(optional) obtained from the user input,
default value set to 5e-6}
}
\description{
binning dependency 2
}
