% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning_2}
\alias{binning_2}
\title{Binning function}
\usage{
binning_2(mass, intensities, samples, tolerance)
}
\arguments{
\item{mass}{}

\item{intensities}{}

\item{samples}{}

\item{tolerance}{}
}
\value{

}
\description{
Binning function
}
