% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binPeaks}
\alias{binPeaks}
\title{Binning of the peaks}
\usage{
binPeaks(df_list, tolerance = 5e-06)
}
\arguments{
\item{df_list}{list of dataframes obtained from the input files}

\item{tolerance}{(optional) obtained from the user input,
default value set to 5e-6}
}
\description{
sorting peaks into bins, relies on binning dependencies 1&2
}
