% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{blank_subtraction}
\alias{blank_subtraction}
\title{Background removal}
\usage{
blank_subtraction(
  dataframe,
  filter_type = "median",
  blank_thresh = 1,
  nsamples_thresh = 1,
  blank_regx = "blank",
  filtered_df = FALSE
)
}
\description{
Background removal
}
