% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{blankSubstraction}
\alias{blankSubstraction}
\title{Substract Blanks from SummarizedExperiment}
\usage{
blankSubstraction(
  exp,
  assay = 1,
  foldChange = 5,
  nSamples = Inf,
  removeBlanks = TRUE
)
}
\arguments{
\item{exp}{SummarizedExperiment obtained after alignment}

\item{foldChange}{Multiplier used for the median blanks. If the Area of
a sample is lower than this threshold, it counts towards the \code{nSamples}
threshold. (Defaults to 5).}

\item{removeBlanks}{Should samples of the type \code{BLANK} be removed after the
filter is applied? Defaults to \code{TRUE}.}

\item{sampleThresh}{Number of samples are allowed to be lower than the
median blank area times the number given for \code{foldChange}. Defaults to \code{Inf},
meaning that no compounds will be removed unless changed.}
}
\value{
SummarizedExperiment with removed compounds and/or samples
according to the parameters set.
}
\description{
This function removes peaks if the signal in the samples is
too low in comparison with the blanks.
}
\details{
For each peak, the median value is taken of the supplied assay and
multiplied with the value in \code{foldChange}. If the number of samples exceeds
the value in \code{sampleThresh}, the signal in the cells is considered too low
and the peak will be removed. By default, \code{sampleThresh} is set to infinity,
so the function does not remove peaks if left unset. Finally, the blank
samples are optionally removed with \code{removeBlanks}
}
\examples{
# Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Substract blanks
blankSubstraction(sumRnegative, nSamples = 4)
}
