% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{bmflAliquots}
\alias{bmflAliquots}
\title{Function specfic for formatting aliquots in BMFL format}
\usage{
bmflAliquots(df)
}
\arguments{
\item{df}{Dataframe of metadata}
}
\description{
Function specfic for formatting aliquots in BMFL format
}
