% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{calculate_nominal_mass}
\alias{calculate_nominal_mass}
\title{Calculate nominal mass from a formula}
\usage{
calculate_nominal_mass(formulas)
}
\arguments{
\item{formulas}{formulas in string format like 'c6h12o6'}
}
\description{
Calculate nominal mass from a formula
}
