% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{cellAlignment}
\alias{cellAlignment}
\title{Align spectra across cells using binning or clustering}
\usage{
cellAlignment(
  spectraDf,
  method = "binning",
  ppm = 5,
  cellData = NULL,
  phenotype = NULL
)
}
\arguments{
\item{spectraDf}{DataFrame of peaks in spectra, where 1 row equals 1 peak
from a single cell}

\item{method}{Character, one of either \code{"binning"} (default) or
\code{"clustering"}.}

\item{ppm}{Tolerance between masses to be considered part of the same bin
or cluster in parts per million (ppm). Defaults to 5.}

\item{cellData}{data.frame of metadata of each cell. While optional in
this step, it is crucial for post-processing. Defaults to \code{NULL}}

\item{phenotype}{Name of the column in \code{cellData} that describes the
phenotype of each cell. Mandatory for statistics and modelling, but can
be left empty here. Defaults to \code{NULL}}
}
\value{
SummarizedExperiment object that contains both data and metadata.
}
\description{
This function will perform either binning (default) or
clustering to group peaks from spectra across cells.
}
\details{
Alignment across cells is a vital part of metabolomics data analysis
in order to make comparisons across cells. Here, we support two methods of
alignment: binning and clustering. Binning is an adjustment from Maldiquant
while clustering is using xcms instead. Both will result in a
SummarizedExperiment object that can be used for postprocessing.
}
\examples{
# Define files
folder <- system.file("cells", package = "sumR")
files <- list.files(folder, full.names = TRUE)

# Extract peaks from files
peaks <- extractPeaks(files)

# Align peaks across spectra
spectra <- spectraAlignment(peaks)

# Align spectra across cells
cells <- cellAlignment(spectra, method = "binning")

## Same but with clustering
# cells <- cellAlignment(spectra, method = "clustering")

# Show output
print(cells)
}
