% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{cellBinning}
\alias{cellBinning}
\title{Bin peaks from spectra across cells.}
\usage{
cellBinning(spectra, ppm = 5)
}
\arguments{
\item{spectra}{DataFrame of spectra obtained from \link{spectraAlignment}.}

\item{ppm}{Tolerance between masses to be considered part of the same bin in
parts per million (ppm). Defaults to 5.}
}
\value{
A SummarizedExperiment with an Area assay and mz, rt and ppm
metrics in rowData.
}
\description{
This function will use the same binning method as
\link{spectraBinning} but across cells instead of spectra.
}
\details{
Binning is a form of alignment that groups together masses
within a given tolerance. Due to measurements errors in mass spectrometry,
binning is necessary to determine which mass intensities belong to
the same peak.
}
