% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPlots.R
\name{cellPlot}
\alias{cellPlot}
\title{plotCellPeaks}
\usage{
cellPlot(peaks, file = 1)
}
\arguments{
\item{peaks}{List of datafres with peak picked data}

\item{file}{File number of the files that were used.}
}
\description{
plotCellPeaks
}
