% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{check_binning}
\alias{check_binning}
\title{Alignment analysis (optional boxplot)}
\usage{
check_binning(
  aligned_peaks,
  summary_errors = F,
  boxplot = F,
  xcoords = c(-50, 0)
)
}
\arguments{
\item{aligned_peaks}{dataframe obtained from `iteration`}

\item{summary_errors}{(optional)logical value obtained from
user input per default set to FALSE}

\item{boxplot}{(optional) logical value obtained from user
input per default set to FALSE}

\item{xcoords}{(optional) vector defining zoom on boxplot
obtained from user input or use of default value c(-50, 0)}
}
\description{
various kinds of analysis are possible,
check_binning ensures m/z values are correctly aligned/binned,
check_binning outputs either a dataframe (1) or
a list of two to three elements:
1- Dataframe of 1 column containing the ppm error values
2- (optional)table of summary stats of ppm error values
3- (optional)- boxplot of the ppm error values
with xcoords zoomed in to -50,0 (default)
}
