% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{check_process}
\alias{check_process}
\title{Checking the results of the alignment with boxplot output if desired}
\usage{
check_process(
  aligned_peaks,
  summary_errors = F,
  boxplot = F,
  xcoords = c(-50, 0)
)
}
\arguments{
\item{aligned_peaks}{dataframe obtained from `iteration`}

\item{summary_errors}{logical value obtained from user input per default set to FALSE}

\item{boxplot}{logical value obtained from user input per default set to FALSE}

\item{xcoords}{vector obtained from user input or use of default value c(-50, 0)}
}
\description{
here the user can choose what kind of analysis they want to have on
their alignment, check_process makes sure that all the m/z values are aligned/binned
correctly, check_process takes a data frame of peaks with mz column as an argument,
and the coordinates for the plot to be zoomed in on, as an optional argument
check_process outputs either a dataframe (1) or a list of two to three elements:
1- Dataframe of 1 column containing the ppm error values
2- (optional)table of summary stats of ppm error values
3- (optional)- boxplot of the ppm erro values with xcoords zoomed in to -50,0 (default)
}
