% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPlots.R
\name{chromatogramPlot}
\alias{chromatogramPlot}
\title{Plot chromatogram of a peak for 1 or more samples}
\usage{
chromatogramPlot(exp, peaks, sampleIdx = 1, peakId = 1)
}
\arguments{
\item{exp}{SummarizedExperiment object obtained after alignment with
\link{cellAlignment}.}

\item{peaks}{List of centroided peaks obtained from \link{extractPeaks}}

\item{sampleIdx}{Vector of 1 or more sample indices or names to plot the
chromatogram for.}

\item{peakId}{Peak index or peak name to plot the chromatogram for.}
}
\description{
This function plots the chromatogram of an aligned peak. This
can be used to inspect the peak alignment.
}
