% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{combine_spectra_centroid}
\alias{combine_spectra_centroid}
\title{Combine spectra and centroid}
\usage{
combine_spectra_centroid(data)
}
\arguments{
\item{data}{XCMS object}
}
\description{
Combine spectra and centroid
}
