% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{condition}
\alias{condition}
\title{Deletion of unwanted samples}
\usage{
condition(mass, intensities, samples, tolerance = 5e-06)
}
\arguments{
\item{mass}{obtained from the input files}

\item{intensities}{obtained from the input files}

\item{samples}{obtained from the input files}

\item{tolerance}{(optional) obtained from the user input,
default value 5e-6}
}
\description{
binning dependency 1
delete any duplication of a samples
delete any sample that is kicked out by the tolerance
}
