% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{constructSE}
\alias{constructSE}
\title{Construct a SummarizedExperiment from aligned spectra}
\usage{
constructSE(res, spectra)
}
\description{
Construct a SummarizedExperiment from aligned spectra
}
