% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{CV_data}
\alias{CV_data}
\title{Cross-validation data}
\usage{
CV_data(data, imp_all, class1, class2)
}
\arguments{
\item{data}{the training set or test set}

\item{imp_all}{the dataframe of all imp m/z from the list of randomForest_CV function}

\item{class1}{sample group 1 name for renaming}

\item{class2}{sample group 2 name for renaming}
}
\description{
Cross-validation data
}
