% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{data_imputation}
\alias{data_imputation}
\title{data imputation}
\usage{
data_imputation(data, noise, seed = 42)
}
\arguments{
\item{data}{dataframe}

\item{noise}{numerical value for the noise level}

\item{seed}{global seed for reproducible results}
}
\description{
This function apply data imputation from 1 to selected noise level
}
