% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{data_split}
\alias{data_split}
\title{datasplit}
\usage{
data_split(seed, data, split_ratio)
}
\arguments{
\item{seed}{global seed}

\item{data}{the dataframe with m/z as columns and a sample column named samples}

\item{split}{ratio a certain value for split ratio}
}
\description{
datasplit
}
