% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{dims_pipeline}
\alias{dims_pipeline}
\title{DIMS pipeline}
\usage{
dims_pipeline(data, plot_md = FALSE)
}
\arguments{
\item{data}{}

\item{plot_md}{}
}
\description{
DIMS pipeline
}
\examples{
dims_pipeline(read_mzml())
}
