% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanFilter.R
\name{doCentroid}
\alias{doCentroid}
\title{Perform centroiding of scans in peaks}
\usage{
doCentroid(peaks, scans = seq_len(length(x)), combineSpectra = FALSE)
}
\arguments{
\item{peaks}{List of dataframes, usually obtained from `peaks()` of the
`mzR` package.}

\item{scans}{Vector of indices, which scans should be used for centroiding?
Defaults to all scans in `peaks`}

\item{combineSpectra}{Boolean value, should scans of different masswindows
be combined into a full-range scan before centroiding? Defaults to `FALSE`}
}
\value{
data.frame of centroided peaks with the columns `scan`, `mz`, `i`
and `Noise`
}
\description{
This function converts profile-mode data into centroided
data. In case of the use of different mass-windows, the `scans` parameter
can be set to subset the scans to be used. The parameter `combineSpectra`
can be set to TRUE if you want to combine the multiple mass-windows into
a single full-range scan.
}
\details{
Centroiding is a major part of a typical metabolomics pipeline.
It reduces the data from a profile-mode dataset drastically, by picking
the local maxima of profiles after smoothing. While in high-resolution
datasets peak picking is followed afterwards, for single-cell DI-MS, peak
picking after centroiding provides sub-optimal results. In many cases, the
peak picking would find peaks originating from the solution, instead of the
compounds in the cells. Therefore, it is recommended to only perform
centroiding and remove the solution-originating peaks afterwards using
filtering.
}
