% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{dunnTest}
\alias{dunnTest}
\title{Dunn test}
\usage{
dunnTest(exp, assay = 1, classifiers = metadata(exp)$phenotype, method = "bh")
}
\arguments{
\item{method}{a character string for correction method .. default is "bh"}
}
\description{
Dunn post hoc test after kruskal-wallis test
}
