% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{extract_features}
\alias{extract_features}
\title{Extract identified features with intensities}
\usage{
extract_features(data, output)
}
\arguments{
\item{data}{readMSData object}

\item{output}{Directory of where the data is stored}
}
\description{
Extract identified features with intensities
}
