% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanFilter.R
\name{extractPeaksBatch}
\alias{extractPeaksBatch}
\title{Format scans in a directory}
\usage{
extractPeaksBatch(directory, ...)
}
\arguments{
\item{...}{see `prepareFile`}
}
\description{
Format scans in a directory
}
