% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{feature_processing}
\alias{feature_processing}
\title{Process features obtained from XCMS}
\usage{
feature_processing(data, impute_method = "knn", remove_blanks = TRUE)
}
\arguments{
\item{data}{XCMS object with grouped peaks}

\item{impute_method}{}

\item{remove_blanks}{Should blank samples be removed?}
}
\description{
Process features obtained from XCMS
}
\examples{
data <- dims_pipeline(read_mzml())
feature_processing(data)
}
