% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPlots.R
\name{featureCovPlot}
\alias{featureCovPlot}
\title{Plot feature coverage when adding cells}
\usage{
featureCovPlot(cells, assay = 1, by = 5, seed = 42)
}
\description{
Plot feature coverage when adding cells
}
