% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{featureFilter}
\alias{featureFilter}
\title{Remove grouped peaks (features) per phenotype}
\usage{
featureFilter(exp, assay = 1, nCells = 0, fCells = 0)
}
\arguments{
\item{exp}{SummarizedExperiment obtained after alignment}

\item{assay}{Assay to be used, defaults to the assay at index 1}

\item{nCells}{Minimum number of cells a feature should be found in. Defaults
to 0, meaning all features are kept.}

\item{fCells}{Similar to \code{nCells} but for fractions instead. Can be used
to remove features that are found in less than half of the cells. Defaults
to 0, meaning all features are kept.}
}
\description{
Remove grouped peaks (features) per phenotype
}
