% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{filterIsotopes}
\alias{filterIsotopes}
\title{Remove identified isotopes from the experiment}
\usage{
filterIsotopes(exp)
}
\arguments{
\item{exp}{SummarizedExperiment with identified isotopes}
}
\description{
Remove identified isotopes from the experiment
}
