% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filterScansByBiomarker}
\alias{filterScansByBiomarker}
\title{Filter scans based on Biomarkers like oil or solvent}
\usage{
filterScansByBiomarker(fileList, mass, intensity = 1e+05, ppm = 20)
}
\arguments{
\item{fileList}{List of centroided peaks obtained from `extractPeaks()`}

\item{mass}{Biomarker mass to use as identification in the solvent}

\item{intensity}{Intensity value that is expected in the solvent}

\item{ppm}{Mass error for the biomarker. Defaults to 20}
}
\description{
Filter scans based on Biomarkers like oil or solvent
}
