% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_func.R
\name{final_glmnet}
\alias{final_glmnet}
\title{final glmnet model}
\usage{
final_glmnet(training, test, alpha, type.measure, seed = NULL, nfolds = 10)
}
\arguments{
\item{training}{training set}

\item{test}{test set}

\item{type.measure}{loss to use for cross-validation. type.measure="class" ,"auc"}

\item{seed}{global seed for reproducible results}

\item{nfolds}{number of folds for cross validation}
}
\description{
This function makes the final model of glmnet using the desired alpha value
}
