% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_func.R
\name{final_glmnet_assess}
\alias{final_glmnet_assess}
\title{final glmnet model assess}
\usage{
final_glmnet_assess(fit_final, test, lambda = "lambda.min")
}
\arguments{
\item{test}{test set}

\item{lambda}{the choice of lambda . default is "lambda.min"}

\item{final_fit}{the cv.glmnet model}
}
\description{
final glmnet model assess
}
