% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_func.R
\name{final_glmnet_features}
\alias{final_glmnet_features}
\title{final glmnet model confusion matrix}
\usage{
final_glmnet_features(fit_final, master_df, lambda = "lambda.min")
}
\arguments{
\item{master_df}{the master df of the raw or scaled intensitites that have m/z column}

\item{lambda}{the choice of lambda . default is "lambda.min"}

\item{final_fit}{the cv.glmnet model}
}
\description{
final glmnet model confusion matrix
}
