% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_func.R
\name{final_glmnet_plot}
\alias{final_glmnet_plot}
\title{final glmnet model plot}
\usage{
final_glmnet_plot(fit_final)
}
\arguments{
\item{final_fit}{the cv.glmnet model}
}
\description{
final glmnet model plot
}
