% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{foldChangeGroups}
\alias{foldChangeGroups}
\title{foldchange for more than two groups test}
\usage{
foldChangeGroups(dataframe2, samples)
}
\arguments{
\item{dataframe2}{transposed dataframe with m/z as columns (raw imputed dataframe)}

\item{samples}{a factor vector with the classes of the samples}
}
\description{
foldchange for more than two groups test
}
