% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{fragmentFilter}
\alias{fragmentFilter}
\title{Identify and remove fragments from aligned peaks}
\usage{
fragmentFilter(exp, assay = 1, method = "spearman", corr = 0.95)
}
\arguments{
\item{exp}{SummarizedExperiment object obtained after alignment}

\item{assay}{Name or index of the assay to use. Defaults to the first
assay (index 1)}

\item{method}{One of either "spearman" or "pearson". It is used as method
in the stats \code{cor} function.}

\item{corr}{Minimum correlation for peaks to be considered fragments.
Defaults to 0.95}
}
\description{
This function finds correlation based on intensities given
in the assay. Peaks that have a higher correlation than the value
given in \code{corr} will be considered fragments and are removed.
}
\details{
Fragments occur in mass spectrometry quite often. Since these
fragments originate from the same compound / metabolite, the resulting
intensities are expected to be very similar. By using either \code{spearman} or
\code{pearson} correlation, the correlation between peaks can be calculated
and be removed.

If a lot of values were originally missing, it is recommended to use
spearman correlation instead of pearson correlation as the latter tends to
give higher correlation values. Spearman correlation uses ranked based
correlation and has shown to give more accurate correlation values when
many values are imputed.
}
