% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{FUN_text_exporter}
\alias{FUN_text_exporter}
\title{Title}
\usage{
FUN_text_exporter(rnamed_cnamed_df, file_path, extension)
}
\arguments{
\item{rnamed_cnamed_df}{data.frame with rownamed mz,colnamed samples}

\item{file_path}{desire output path, should be a string}

\item{extension}{desired custom extension to files (aside from .txt), string}
}
\value{

}
\description{
takes teh df, separates it into mz,int files for each sample (1st loop)
export each file with desired file path,and custom extension (2nd loop)
UNALIGNED ! UNLIMITED POWAAAAAH
}
