% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{generateModel}
\alias{generateModel}
\title{Model using RandomForest model}
\usage{
generateModel(
  exp,
  modelName,
  classifiers = metadata(exp)$phenotype,
  assay = 1,
  folds = 5,
  ratio = 0.632,
  seed = NULL,
  ...
)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{classifiers}{Column name in the colData that represents the classifiers}

\item{assay}{Which assay should be used to create a model}

\item{ratio}{Train-test split ratio}

\item{...}{Any other parameters for preProcessing}

\item{cv}{Number of folds to be used for cross validation}
}
\description{
Model using RandomForest model
}
