% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_calibrations}
\alias{get_calibrations}
\title{get vector of Calibration MZs}
\usage{
get_calibrations(mzs)
}
\arguments{
\item{mzs}{Vector of experimental measured M/Zs}
}
\value{
a vector of 'correct' MZ values, used to correct the experimental M/Zs
}
\description{
get vector of Calibration MZs
}
