% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDefect.R
\name{get_inclusion_list}
\alias{get_inclusion_list}
\title{Retrieve inclusion list in a data frame format (default or user input)}
\usage{
get_inclusion_list(file = NULL)
}
\arguments{
\item{file}{(optional) String location of the inclusion list}
}
\description{
Retrieve inclusion list in a data frame format (default or user input)
}
