% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_func.R
\name{glmnet_cv_alphachoice}
\alias{glmnet_cv_alphachoice}
\title{Alpha value choice}
\usage{
glmnet_cv_alphachoice(
  training,
  test,
  type.measure,
  seed,
  lambda = c("lambda.min", "lambda.1se")
)
}
\arguments{
\item{training}{training set}

\item{test}{test set}

\item{type.measure}{loss to use for cross-validation. type.measure="class" ,"auc"}

\item{seed}{global seed for reproducible results}

\item{lambda}{choose between "lambda.min" or "lambda.1se"}
}
\description{
This function choose the alpha value for elastic net regression with the desired accuracy
}
