% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatMap}
\alias{heatMap}
\title{heatmap}
\usage{
heatMap(data, classifiers, pretty.order.rows = T, pretty.order.cols = T)
}
\arguments{
\item{data}{dataframe with m/z as columns}

\item{classifiers}{samples group as factor}

\item{pretty.order.rows}{logical vector the default is TRUE}

\item{pretty.order.cols}{logical vector the default is TRUE}
}
\description{
heatmap plot of m/z intensities of grouped samples
}
