% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{imp_data}
\alias{imp_data}
\title{feature selection dataframe}
\usage{
imp_data(feature_select_model, data, class1, class2)
}
\arguments{
\item{feature_select_model}{feature select model from imp_select function}

\item{data}{the master dataframe}

\item{class1}{sample group 1 name for renaming}

\item{class2}{sample group 2 name for renaming}
}
\description{
feature selection dataframe
}
