% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Impute missing values of an assay}
\usage{
imputation(
  exp,
  method = "noise",
  useAssay = "Area",
  saveAssay = "Imputed",
  setDefault = TRUE,
  ...
)
}
\arguments{
\item{exp}{SummarizedExperiment object obtained after alignment, containing
at least the slot given in \code{useAssay}.}

\item{method}{Character vector of method to use for imputation. Can be one
of \code{"noise"}, \code{"model"}, \code{"magic"}, \code{"saver"}. Defaults to \code{"noise"}.}

\item{useAssay}{Character name of the assay to be imputed. Defaults to
\code{"Area"}}

\item{saveAssay}{Character name of the to be saved imputed assay. Defaults
to \code{"Imputed"}}

\item{setDefault}{Boolean value, should the imputed assay be set at index 1,
making it the default wher using \code{assay(exp)}? Defaults to \code{TRUE}}

\item{...}{
  Arguments passed on to \code{\link[=noiseImputation]{noiseImputation}}, \code{\link[=modelImputation]{modelImputation}}, \code{\link[=saverImputation]{saverImputation}}
  \describe{
    \item{\code{noise}}{numerical value for the maximum random noise level. Defaults
to 100.}
    \item{\code{model}}{Character, name of the model to be used for missing value
imputation. Requires the \code{pmp} packages and must be one of the
methods mentioned in \code{mv_imputation}.}
    \item{\code{normalize}}{Boolean value, should data be normalized before imputation?}
  }}
}
\value{
A SummarizedExperiment object with an added slot in assays given
with the \code{saveAssay} parameter.
}
\description{
This function aids in imputation of peaks that were not found
with \link{extractPeaks}. Currently, four methods are supported of imputation
are supported through the \code{method} parameter. These are based on
random (noise) or model imputation (\link{modelImputation}, \link{magicImputation},
and \link{saverImputation}). The former uses regular models like KNN or Random
Forest models to impute values. The latter two are based on single-cell RNA
methods that are more complex and take correlation between cells into
account.
}
\details{
Imputation of missing values is a vital part of post-processing
as it is needed for other post-processing methods, including statistics and
modelling.
}
\examples{
# Loading negative polarity dataset
data("sumRnegative")

# Noise Imputation
exp <- imputation(sumRnegative, method = "noise", noise = 250)

# Model Imputation
exp <- imputation(sumRnegative, method = "model", model = "sv")

# Rmagic Imputation
# exp <- imputation(sumRnegative, method = "magic")

# SAVER Imputation
exp <- imputation(sumRnegative, method = "saver", normalize = TRUE)
}
