% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{impute.KNN.obs.sel}
\alias{impute.KNN.obs.sel}
\title{Imputation method for missing intensity values}
\usage{
impute.KNN.obs.sel(dat, cor.var.sel = 0.2, K = 5, verbose = T)
}
\description{
Imputation method for missing intensity values
}
