% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{Intensity_ratio_range}
\alias{Intensity_ratio_range}
\title{Intensity ratio for carbon}
\usage{
Alkyne_n(n)
}
\arguments{
\item{mz}{Mass of interest in m/z}

\item{by_step}{"by" in sequence range}
}
\description{
Calculates the intensity ratio for carbon C(n)H(2n-2) within a specified range
}
