% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{iso_hippo}
\alias{iso_hippo}
\title{Iso hippo}
\usage{
iso_hippo(df, isotope_da = 1.0034, iso_tolerance = 0.0034)
}
\arguments{
\item{df}{is the dataframe on which hippo eats, it has to have an mz column}

\item{isotope_da}{is an optional argument, specifying the differences
between mol ions and isotopes in Da}

\item{iso_tolerance}{optional argument: tolerance in Da for isotope differences}
}
\description{
iso_hippo is a function that takes 1 argument, and 2 optional ones
Similarly to kopi luwak, iso hippo is a retired, senile, hippopotamus
that grazes on huge chunks of data and poops isotopes and mol.ions

Ahmed's notes: Should expand on it and allow it to work with other isotopes +
merge it with the repeating units function, also have it check intensity values
#to make sure if its actually an isotope of the parent molecule or not
}
\examples{


}
