% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{isotope_molecules}
\alias{isotope_molecules}
\title{Annotation list for isotopes}
\usage{
isotope_molecules(adducts_isotope, ppm , z )
}
\arguments{
\item{adducts_isotope}{dataframe filtered for isotopes with possible isotopes}

\item{ppm}{error part per million}

\item{z}{charge values}
}
\description{
A list of annotation information of isotopes
}
