% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{isotope_tagging}
\alias{isotope_tagging}
\title{Isotope tagging}
\usage{
isotope_tagging(df , ppm ,z)
}
\arguments{
\item{df}{Dataframe must contain columns c("mz", "intensity") with "mz" as first column}

\item{ppm}{Parts Per Million Tolerance}

\item{z}{charge}
}
\description{
Tags isotopes based on intensity ratio and difference in mass
}
