% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{isotope_valid}
\alias{isotope_valid}
\title{Filtering by validity}
\usage{
isotope_valid(isotope_molecules_list)
}
\arguments{
\item{isotope_molecules_list}{list of annotate information for isotopes}
}
\description{
A list of annotation information of isotopes
}
