% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{isotopeFinding}
\alias{isotopeFinding}
\title{Isotope finding}
\usage{
isotopeFinding(data, mz_vector, isotope_df, Elements = c("C13"), ppm = 5, z = 0, max_diff_mat, min_diff_mat)
}
\arguments{
\item{data}{data.frame, with min_error and max_error columns added}

\item{mz_vector}{vector of mz values}

\item{isotope_df}{data.frame containing isotopic elements of interest, PSE notation and mass difference compared to mono-isotope}

\item{Elements}{vector, with selected isotopic elements of interest (default = c("C13")), options : c("C13","Cl37")}

\item{ppm}{An integer, defining parts per million (ppm) for tolerance (default = 5)}

\item{z}{An integer, defining charge z of m/z peaks for calculation of real mass. 0 is for auto-detection (default = 0)}

\item{max_diff_mat}{matrix, with maximum difference between mz pairs}

\item{min_diff_mat}{matrix, with maximum difference between mz pairs}
}
\description{
Searches for isotopes
}
