% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{isotopeMolecules}
\alias{isotopeMolecules}
\title{List of isotopes with annotation information}
\usage{
isotopeMolecules(adducts_matches_int, Elements = c("C13"), ppm = 5 , z = 0)
}
\arguments{
\item{adducts_matches_int}{data.frame with containing mass and intensity of possible mono-isotopic ion and isotope pairs}

\item{Elements}{vector, with selected isotopic elements of interest (default = c("C13")), options : c("C13","Cl37")}

\item{ppm}{An integer, defining parts per million (ppm) for tolerance (default = 5)}

\item{z}{An integer, defining charge z of m/z peaks for calculation of real mass. 0 is for auto-detection (default = 0)}
}
\description{
Creates a list of isotopes with annotation information
}
