% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{isotopeValid}
\alias{isotopeValid}
\title{Filtering isotopes by validity}
\usage{
isotopeValid(isotope_molecules_list)
}
\arguments{
\item{isotope_molecules_list}{A list, of possible isotopes containing annotate information}
}
\description{
Filtering a list of possible isotopes based on validity
}
