% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{iteration}
\alias{iteration}
\title{Binning iterations}
\usage{
iteration(df_list, max_align = 8, bin_plot = F)
}
\arguments{
\item{df_list}{list of dataframes obtained from `binPeaks`}

\item{max_align}{(optional) value obtained from user input,
default value set to 8}

\item{bin_plot}{(optional) logical value deciding if plot
is created obtained from user input, default set to FALSE,
line plot shows reduction of bins per iteration}
}
\description{
the code iterates the binprak function
over the data as long as the number of bins changes,
when the number of bins doesn't change anymore
iterations stop, maximal number of iterations is
set to 8
plot of the bin decrease can be created (optional)
}
