% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{keepVariableFeatures}
\alias{keepVariableFeatures}
\title{Keep the most variable features based on a univariate test}
\usage{
keepVariableFeatures(
  exp,
  test = "leveneTest",
  threshold = 0.05,
  method = "any"
)
}
\arguments{
\item{exp}{SummarizedExperiment object after alignment and leveneTest}
}
\value{
A SummarizedExperiment with only compounds that have different
variances across the phenotype given.
}
\description{
This function removes features that have similar features
across the given phenotypes and calculated by the function \code{leveneTest}.
}
\details{
When interested in which peaks are significantly different between
groups in the given phenotypes, the leveneTest can be used. Here, only
features are kept where the variance between the groups is significantly
different. This greatly reduces the dataset with often only a few compounds
remaining.
}
