% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{keepVariablePeaks}
\alias{keepVariablePeaks}
\title{Keep the most variable peaks based on a univariate test}
\usage{
keepVariablePeaks(exp, test, threshold = 0.05, method = "any")
}
\arguments{
\item{exp}{SummarizedExperiment object after performing statistical test(s)}

\item{test}{Name of the test executed and stored as a column in \link{rowData}.}

\item{threshold}{Numerical value, p-value threshold for determining
significance after multiple test correction. Defaults to 0.05}

\item{method}{One of \code{"any"} or \code{"all"}. Only used for when number of
phenotypes > 2. If \code{"any"}, any p-value value between two groups below
\code{threshold} will be kept. If \code{"all"}, all p-values between the groups must
be below \code{threshold} in order to be kept. Defaults to \code{"any"}.}
}
\value{
A SummarizedExperiment with only compounds that have different
variances across the phenotype given.
}
\description{
This function removes peaks that have similar peaks
across the given phenotypes and calculated by the function \code{leveneTest}.
}
\details{
When interested in which peaks are significantly different between
groups in the given phenotypes, the leveneTest can be used. Here, only
peaks are kept where the variance between the groups is significantly
different. This greatly reduces the dataset with often only a few compounds
remaining.
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Perform wilcoxText
sumRnegative <- wilcoxTest(sumRnegative)

# Only keep peaks with p-value < 0.1
keepVariablePeaks(sumRnegative, test = "wilcoxTest", threshold = 0.1)
}
