% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{leveneTest}
\alias{leveneTest}
\title{Levene test for unequal variances across phenotypes}
\usage{
leveneTest(
  exp,
  assay = 1,
  classifiers = metadata(exp)$phenotype,
  method = "fdr"
)
}
\description{
levene test for variance check
}
