% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{leveneTestvalues}
\alias{leveneTestvalues}
\title{levene test}
\usage{
leveneTestvalues(dataframe, classifiers, threshold = 0.05)
}
\arguments{
\item{dataframe}{transposed dataframe with m/z as columns}

\item{classifiers}{a factor vector with the classes of the samples}

\item{threshold}{numerical value for wanted threshold . the default is 0.05}
}
\description{
levene test for variance check
}
