% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{magicImputation}
\alias{magicImputation}
\title{Imputation of missing values using Rmagic}
\usage{
magicImputation(data)
}
\arguments{
\item{data}{Data.frame of the assay to be imputated}
}
\value{
A data.frame with all values filled with the same dimensions
as the given argument in \code{data}.
}
\description{
This function aids in imputation of peaks that were not found
with \link{extractPeaks}. Currently, four methods are supported of imputation
are supported through the \code{method} parameter. These are based on
random (noise) or model imputation (\link{modelImputation}, \link{magicImputation},
and \link{saverImputation}). The former uses regular models like KNN or Random
Forest models to impute values. The latter two are based on single-cell RNA
methods that are more complex and take correlation between cells into
account.
}
\details{
Imputation of missing values is a vital part of post-processing
as it is needed for other post-processing methods, including statistics and
modelling.
}
