% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDefect.R
\name{make_filter_list}
\alias{make_filter_list}
\title{Defining MD cut-off}
\usage{
make_filter_list(filter_df)
}
\arguments{
\item{filter_df}{obtained from the HMDB using \code{mass_defect_calculation}}
}
\description{
the theoretical maximum MD of human metabolites
is used to calculate this cut off
}
