% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{mass_spec_plot}
\alias{mass_spec_plot}
\title{Mass spectrometer visualization}
\usage{
mass_spec_plot(final_df_vis)
}
\arguments{
\item{final_df_vis}{dataframe of the mono-isotopic and isotopic peaks pair's with id column}
}
\description{
Gives a graphical representation of the mono-isotopic and isotopic peaks
}
