% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{massSpecPlot}
\alias{massSpecPlot}
\title{Mass spectrometry visualization}
\usage{
massSpecPlot(final_df)
}
\arguments{
\item{final_df}{a data.frame of the mono-isotopic and isotopic peaks pair's with id column}
}
\description{
Gives a interactive graphical representation of the mono-isotopic and isotopic peaks
}
