% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_mz.R
\name{match_mz_with_database}
\alias{match_mz_with_database}
\title{Title}
\usage{
match_mz_with_database(
  df_raw,
  database_dir,
  base.dbname = "hmdb",
  ppm = 5,
  ionmode = c("positive")
)
}
\arguments{
\item{base.dbname}{name of the database that should be used, e.g. hmdb, lipidmaps,}

\item{databaseDir}{path to the folder containing the downloaded database.
See example on how to download one.}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
## instead of a temporary dir, define a fixed path
database_dir = tempdir()
dbname = "lipidmaps"
buildBaseDB(outfolder = database_dir, dbname = dbname, test = FALSE)

matches <- match_mz_with_database(df_raw = df_raw, database_dir = database_dir,
base.dbname = dbname, ppm = 5, ionmode = c("positive"))
}
}
