% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{MDS_plot}
\alias{MDS_plot}
\title{MDS plot}
\usage{
MDS_plot(model, training_set)
}
\arguments{
\item{model}{random forest model}

\item{training_set}{training set that was used for the random forest model with samples column}
}
\description{
random forest plots
}
