% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{metadataFromAliquots}
\alias{metadataFromAliquots}
\title{Receive metadata from aliquot names}
\usage{
metadataFromAliquots(
  aliquots,
  regex = "([0-9]{4}[A-Z]{3}_[0-9]{4})([A-Z]+|)([0-9]+|)_([A-Z])([0-9])",
  type = "BMFL",
  columns = NULL
)
}
\description{
Receive metadata from aliquot names
}
