% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadataFromFile}
\alias{metadataFromFile}
\title{Obtain metadata from aliquot names}
\usage{
metadataFromFile(
  files,
  regex = "([0-9]{4}[A-Z]{3}_[0-9]{4})([A-Z]+|)([0-9]+|)_([A-Z])([0-9])"
)
}
\description{
Obtain metadata from aliquot names
}
