% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model<-}
\alias{model<-}
\title{Set or replace a model in the supplied SummarizedExperiment}
\usage{
model(exp, modelName) <- value
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{modelName}{Name of the model to adjust}

\item{value}{Value to set in the model}
}
\value{
Updated SummarizedExperiment object with added model information
stored in the \link{metadata} slot.
}
\description{
This function aids in setting a model in the \link{metadata} slot
of the SummarizedExperiment if none is available.
}
\details{
This function is used by \link{generateModel} for setting a trained
model and its results.
}
