% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{model}
\alias{model}
\title{Get model in Experiment}
\usage{
model(exp, modelName = 1)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{modelName}{Name of the model to retrieve}
}
\description{
Get model in Experiment
}
