% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_func.R
\name{modelGLM}
\alias{modelGLM}
\title{Model glmnet}
\usage{
modelGLM(
  exp,
  classifiers = metadata(exp)$phenotype,
  assay = 1,
  type.measure = "deviance",
  lambda = c("lambda.min", "lambda.1se"),
  cv = 5,
  ratio = 0.8,
  seed = NULL,
  alphas = seq(0, 1, 0.1)
)
}
\description{
Model glmnet
}
