% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{modelRF}
\alias{modelRF}
\title{Model using RandomForest model}
\usage{
modelRF(
  exp,
  classifiers = metadata(exp)$phenotype,
  assay = 1,
  cv = 5,
  ratio = 0.8
)
}
\description{
Model using RandomForest model
}
