% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{models}
\alias{models}
\title{Get names of models in Experiment}
\usage{
models(exp)
}
\arguments{
\item{exp}{SummarizedExperiment object}
}
\value{
Character vector of the names of models created with
\link{generateModel}.
}
\description{
After modelling, the results are stored in the \link{metadata} slot
as a list with name \code{model}. However, multiple models can be made and stored
this way. This function returns the names of models that have been created.
}
\details{
The \code{model} name is a reserved name in sumR, as all models are
stored in a list with the name \code{model}. Created models and their results like
cross-validation and variable importances are stored in a sub-list of
\code{model}. These can be accessed with the \link{model} function, which requires
a model name or index. This function aids in this process by returning the
names of all models that have been created with \link{generateModel}.
}
\examples{
# Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Generate a (Random Forest) model
sumRnegative <- generateModel(sumRnegative, "rf", ratio = 0.5)

# Retrieve model infomation
model(sumRnegative)

# Retrieve Variable importances
varImportance(sumRnegative, "rf")
}
