% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{mtry_select}
\alias{mtry_select}
\title{choose mtry}
\usage{
mtry_select(training_set, test_set, seed, ntree)
}
\arguments{
\item{training_set}{the training set}

\item{test_set}{the test set}

\item{seed}{global seed for reproducible results}

\item{ntree}{number of trees used in the original model}
}
\description{
This function choose the mtry according to the user preference (Accuracy, Sensitivity, Specificity)
}
