% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{mydensityClust_1}
\alias{mydensityClust_1}
\title{Mydensity clust 1}
\usage{
mydensityClust_1(data, mzMin, mzMax, dc)
}
\arguments{
\item{data}{}

\item{mzMin}{}

\item{mzMax}{}

\item{dc}{}
}
\description{
Mydensity clust 1
}
