% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_plot.R
\name{mz_analysis_volcano_magic}
\alias{mz_analysis_volcano_magic}
\title{MZLOG-OBJ Volcano Magic
Create Welch, Fold & Volcano plot, with default inputs}
\usage{
mz_analysis_volcano_magic(input_mzlog_obj, phenotype_a, phenotype_b, cores = 2)
}
\arguments{
\item{input_mzlog_obj}{\cr
DataFrame : Log2 of Input MZ-Obj}

\item{phenotype_a}{\cr
DataFrame: Filtered metadata for phenotypeA}

\item{phenotype_b}{\cr
DataFrame: Filtered metadata for phenotypeB}

\item{cores}{Integer: Can I has multithreading? (Need parallel)}
}
\description{
MZLOG-OBJ Volcano Magic
Create Welch, Fold & Volcano plot, with default inputs
}
