% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_mz_post_processing.R
\name{mz_pp_imputation}
\alias{mz_pp_imputation}
\title{MZ-OBJ Imputation}
\usage{
mz_pp_imputation(input_mz_obj, low_noise = 10, high_noise = 5000)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{low_noise}{\cr
Float     : LowBoundry for "Noise"}

\item{high_noise}{\cr
Float     : HighBoundry for "Noise"}
}
\value{
MZ-OBJ
}
\description{
Replace all <\link{low_noise} ("mostly zeros") with a randomized value
between \link{low_noise} and \link{high_noise}
}
