% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_mz_post_processing.R
\name{mz_pp_normalization}
\alias{mz_pp_normalization}
\title{MZ-OBJ Normalization}
\usage{
mz_pp_normalization(input_mz_obj, metadata, plot = TRUE)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{plot}{\cr
Boolean   : to plot or not to plot}
}
\value{
Returns an MZ-OBJ
}
\description{
Not really sure what this does
\itemize{
\item Requires: ggplot2, dplyr
}
}
