% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_mz_quality_check.R
\name{mz_quality_magic}
\alias{mz_quality_magic}
\title{MZ-OBJ Quality Magic}
\usage{
mz_quality_magic(input_mz_obj, meta, cores = 2)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{meta}{\cr
DataFrame : Input MZ-Obj}

\item{cores}{Integer: Can I has multithreading? (Need parallel)

Dependencies : dplyr}
}
\value{
Returns a dataframe of samples by metrics
}
\description{
MZ-OBJ Quality Magic
}
