% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_mz_quality_check.R
\name{mz_quality_meta_check}
\alias{mz_quality_meta_check}
\title{Does the metadata seem ok}
\usage{
mz_quality_meta_check(input_mz_obj, meta)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{meta}{\cr
DataFrame : Input MZ-Obj

Dependencies : dplyr}
}
\description{
Does the metadata seem ok
}
