% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_6_analysis.R
\name{mz_random_forest}
\alias{mz_random_forest}
\title{MZ-OBJ RandomForest}
\usage{
mz_random_forest(input_mz_obj)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{metadata}{\cr
DataFrame?: MZ_metadata}

\item{bad_samples}{\cr
List?     : c("bad1", "bad2")}

\item{high_noise}{\cr
Float     : HighBoundry for "Noise"}
}
\description{
Not really sure
\itemize{
\item Requires: ggplot2, tibble
}
}
