% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mzlog_analysis.R
\name{mzlog_analysis_fold}
\alias{mzlog_analysis_fold}
\title{MZLOG-OBJ Fold Change}
\usage{
mzlog_analysis_fold(input_mzlog_obj, phenotype_a, phenotype_b, cores = 2)
}
\arguments{
\item{input_mzlog_obj}{\cr
DataFrame : Log2 of Input MZ-Obj}

\item{phenotype_a}{\cr
DataFrame: Filtered metadata for phenotypeA}

\item{phenotype_b}{\cr
DataFrame: Filtered metadata for phenotypeB}

\item{cores}{Integer: Can I has multithreading? (Need parallel)}
}
\value{
DataFrame with columns:
\itemize{
\item mz   : Float
\item fold : Float
}
}
\description{
Compare two phenotypes
}
