% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mzlog_analysis.R
\name{mzlog_analysis_pca}
\alias{mzlog_analysis_pca}
\title{MZLOG-OBJ PCA}
\usage{
mzlog_analysis_pca(input_mzlog_obj, metadata, bad_samples)
}
\arguments{
\item{input_mzlog_obj}{\cr
DataFrame : Log2 of Input MZ-Obj}

\item{metadata}{\cr
DataFrame: MZ_metadata}

\item{bad_samples}{\cr
List?     : c("bad1", "bad2")}
}
\description{
Not really sure
\itemize{
\item Requires: ggplot2, tibble
}
}
