% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mzlog_randomforest.R
\name{mzlog_rf_correlation}
\alias{mzlog_rf_correlation}
\title{Find Correlation data within a mzLog_obj}
\usage{
mzlog_rf_correlation(input_mzlog_obj, correlation_cutoff = 0.75)
}
\arguments{
\item{input_mzlog_obj}{\cr
DataFrame : Log2 of Input MZ-Obj}

\item{correlation_cutoff}{\cr
Float: Decimal percentage of high correlation cutoff}
}
\value{
Transposed mzlog_obj
}
\description{
Find Correlation data within a mzLog_obj
}
