% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mzlog_analysis.R
\name{mzlog_welch}
\alias{mzlog_welch}
\title{MZ-OBJ T-Test}
\usage{
mzlog_welch(input_mzlog_obj, phenotype_a, phenotype_b, cores = 4)
}
\arguments{
\item{input_mzlog_obj}{\cr
DataFrame : Log2 of Input MZ-Obj}

\item{phenotype_a}{\cr
DataFrame: Filtered metadata for phenotypeA}

\item{phenotype_b}{\cr
DataFrame: Filtered metadata for phenotypeB}
}
\value{
DataFrame with columns:
\itemize{
\item rowname: Float: MZ
\item p      : Float: p-value
\item p_05   : Boolean: is P < 0.05
\item p_10   : Boolean: is P < 0.10
\item p_15   : Boolean: is P < 0.15
}
}
\description{
Welch T-test, input should be an log2 of mz_obj
\itemize{
\item Requires: dplyr
'
}
}
