% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_mzlong_postprocessing.R
\name{mzlong_pp_log_transform}
\alias{mzlong_pp_log_transform}
\title{MZLONG-OBJ Log Transform}
\usage{
mzlong_pp_log_transform(input_mzlong, plot = TRUE)
}
\arguments{
\item{input_mzlong}{\cr
DataFrame : Input MZLONG-Obj}

\item{plot}{\cr
Boolean   : to plot or not to plot}
}
\value{
mzLong_obj
}
\description{
Convert intensities of MZLong to log2
For normal "mz_obj" use log2(mz_obj)
\itemize{
\item Requires: ggplot2
}
}
