% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_mz_quality_check.R
\name{mzmetrics_quality_plot}
\alias{mzmetrics_quality_plot}
\title{MZ_METRICS plot a single statistic}
\usage{
mzmetrics_quality_plot(mz_metrics, focus, title = focus)
}
\arguments{
\item{mz_metrics}{\cr
DataFrame : from "sumR::mz_quality_metrics"}

\item{focus}{\cr
String : which metric to plot (i.e. median_mz)}

\item{title}{\cr
String : title of plot

Dependencies : ggplot2, dplyr, parallel}
}
\description{
MZ_METRICS plot a single statistic
}
