% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_mz_quality_check.R
\name{mzmetrics_quality_plot_all}
\alias{mzmetrics_quality_plot_all}
\title{MZ_METRICS plot all statistics}
\usage{
mzmetrics_quality_plot_all(mz_metrics)
}
\arguments{
\item{mz_metrics}{\cr
DataFrame : from "sumR::mz_quality_metrics"}
}
\description{
MZ_METRICS plot all statistics
}
