% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_1_extract_centroid.R
\name{mzT_filtering}
\alias{mzT_filtering}
\title{MzTime filtering (Uses mz_filtering) & binning for an MzT obj.}
\usage{
mzT_filtering(
  mzT,
  method = max,
  missingness_threshold = 1,
  intensity_threshold = 1000
)
}
\arguments{
\item{mzT}{\cr
MzT : MzT object}

\item{method}{\cr
\link{Math} : i.e. (mean, max, median)}

\item{missingness_threshold}{\cr
Numeric   : If Decimal : Threshold \% to blank data under. (as a decimal 10\% = 0.1)
: If Whole : Number of scans required to be nonzero ( Value required in at least 2 scans )}

\item{intensity_threshold}{\cr
Numeric   : Lowest allowed intensity}
}
\description{
MzTime filtering (Uses mz_filtering) & binning for an MzT obj.
}
