% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{noiseImputation}
\alias{noiseImputation}
\title{data imputation}
\usage{
noiseImputation(data, noise = 100, seed = 42)
}
\arguments{
\item{data}{dataframe}

\item{noise}{numerical value for the noise level}

\item{seed}{global seed for reproducible results}
}
\description{
This function apply data imputation from 1 to selected noise level
}
