% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{normalize_features}
\alias{normalize_features}
\title{Normalize intensity dataframe}
\usage{
normalize_features(intensity_df)
}
\arguments{
\item{intensity_df}{Dataframe with intensities. Can contain isotope information}
}
\description{
Normalize intensity dataframe
}
