% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parallel_coord}
\alias{parallel_coord}
\title{parallel coordinates plot}
\usage{
parallel_coord(
  data,
  tag = c(TRUE, FALSE),
  columns,
  groupColumn,
  scale,
  boxplot = FALSE,
  highlight = highlight,
  highlight_color = c("maroon", "navy")
)
}
\arguments{
\item{data}{dataframe}

\item{tag}{logical vector TRUE or FALSE if highlighting a certain value is needed or not}

\item{columns}{the columns index for plotting either intensity or median intensity .. etc}

\item{groupColumn}{the column name of the groups or the index of the group column}

\item{scale}{character vector specifying the type of scaling "globalminmax","uniminmax","std","robust"}

\item{boxplot}{logical vector to add a boxplot to the plot, the default is FALSE}

\item{highlight}{if tag=TRUE, add highlighting ifelse statement:
example : highlight=if_else(data$ppm_difference < 5, "ppm difference less than 5", "ppm difference more than 5")
example : highlight=if_else(data$class1 > data$class2 ,"class1","class2")}

\item{highlight_color}{the highlight color, the default is c("maroon","navy")}
}
\description{
parallel coordinates plot with either highlighting a certain value or not
}
