% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroiding.R
\name{parseFile}
\alias{parseFile}
\title{Parse a provided mzML file}
\usage{
parseFile(
  file,
  massWindow = c(0, Inf),
  polarity = "-",
  combineSpectra = FALSE,
  cl = NULL
)
}
\arguments{
\item{file}{A mzML file in profile-mode}

\item{massWindow}{A vector of size 2 with min-max size of mass-windows to be
used. E.g. A mass-window from 200 to 300 has a mass-window of 100. Defaults
to c(0, Inf) indicating that all mass-windows should be taken into account}

\item{polarity}{A single character of either "-" (negative) or "+"
(positive) indicating the polarity to be used. Defaults to "-"}

\item{combineSpectra}{Boolean value, should scans of different masswindows
be combined into a full-range scan before centroiding? Defaults to `FALSE`}
}
\value{
data.frame of centroided peaks with the columns `scan`, `mz`, `i`
and `Noise`. It also stores the used parameters as attributes of the
data.frame
}
\description{
Centroiding is a vital part of a typical metabolomics workflow.
This function centroids a given mzML file and returns a dataframe of
centroided peaks. See `doCentroid` for more information about centroiding.
}
\details{
Centroiding is a major part of a typical metabolomics pipeline.
It reduces the data from a profile-mode dataset drastically, by picking
the local maxima of profiles after smoothing. While in high-resolution
datasets peak picking is followed afterwards, for single-cell DI-MS, peak
picking after centroiding provides sub-optimal results. In many cases, the
peak picking would find peaks originating from the solution, instead of the
compounds in the cells. Therefore, it is recommended to only perform
centroiding and remove the solution-originating peaks afterwards using
filtering.
}
