% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PCA_ellipse}
\alias{PCA_ellipse}
\title{PCA ellipse plot}
\usage{
PCA_ellipse(data, classifiers)
}
\arguments{
\item{data}{transposed dataframe with m/z as columns}

\item{classifiers}{sample groups as factor}
}
\description{
PCA ellipse plot
}
