% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{peakFilter}
\alias{peakFilter}
\title{Remove grouped peaks per phenotype}
\usage{
peakFilter(
  exp,
  assay = 1,
  phenotype = metadata(exp)$phenotype,
  nCells = 0,
  fCells = 0
)
}
\arguments{
\item{exp}{SummarizedExperiment obtained after alignment}

\item{assay}{Assay to be used, defaults to the assay at index 1}

\item{phenotype}{Name of the column used as a phenotype. Defaults to the
phenotype in the \link{metadata} slot, set by \link{setPhenotype}}

\item{nCells}{Minimum number of cells a feature should be found in. Defaults
to 0, meaning all features are kept.}

\item{fCells}{Similar to \code{nCells} but for fractions instead. Can be used
to remove features that are found in less than half of the cells. Defaults
to 0, meaning all features are kept.}
}
\value{
SummarizedExperiment with removed peaks determined by the
parameters provided.
}
\description{
This function removes peaks that might be discovered in only
a few cases per phenotype and therefore undesired for statistics and
modelling.
}
\details{
Per phenotype in the set phenotype column of \link{colData}, this
function filters peaks by either absolute (\code{nCells}) or by fraction
(\code{fCells}).

In most cases, peaks often are discovered in just a single cell due to
technical variation. By setting \code{nCells = 2}, these peaks are removed.
Likewise, to only keep the peaks found in half of the phenotypes,
not considering the number of samples, set \code{fCells = 0.5}. A combination of
the two might also be a valuable strategy, but should be considered wisely.
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Filter peaks based on absolute numbers
peakFilter(sumRnegative, nCells = 3)

# Filter peaks based on fractions
peakFilter(sumRnegative, fCells = 0.5)
}
