% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{peakGrouping}
\alias{peakGrouping}
\title{Group the peaks}
\usage{
peakGrouping(
  l,
  fraction = 0,
  npeaks = 0,
  cores = detectCores(logical = F),
  meanSNR = 0
)
}
\description{
Group the peaks
}
