% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{peakPickFile}
\alias{peakPickFile}
\title{Perform peak picking of a single file}
\usage{
peakPickFile(
  file = NULL,
  scales = c(1, 4, 9),
  noiseWindow = 0.1,
  spectra = NULL,
  ...
)
}
\description{
Perform peak picking of a single file
}
