% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{Perform peak picking}
\usage{
peakPicking(
  files,
  polarity = NULL,
  scaleFactor = 1,
  combineSpectra = FALSE,
  massWindow = c(0, Inf),
  noiseWindow = 0.1,
  scales = NULL,
  cores = detectCores(logical = F),
  solventFilter = NULL
)
}
\description{
Perform peak picking
}
