% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{permutation.test}
\alias{permutation.test}
\title{data permutation}
\usage{
permutation.test(training_set, test_set, model, n, class1, class2, sample_no)
}
\arguments{
\item{training_set}{the training set}

\item{test_set}{the test set}

\item{model}{the random forest model}

\item{n}{number permutation}

\item{class1}{sample group 1 name for renaming}

\item{class2}{sample group 2 name for renaming}

\item{sample_no}{the number of all samples(training+test set)}

\item{seed}{global seed for reproducible results}
}
\description{
This function apply data permutation to test the accuracy of the model against permutated data
}
