% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pheatmapPlot}
\alias{pheatmapPlot}
\title{Heatmap with colData groups}
\usage{
pheatmapPlot(exp, assay = 1, columns = NULL, rows = NULL, ...)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{assay}{Name of the assay or index of the assay to use}

\item{columns}{Vector of column names from the colData to use as
groups in the heatmap}

\item{...}{Any other parameters submitted to the \code{pheatmap} function}
}
\description{
Heatmap with colData groups
}
