% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{pickSpectra}
\alias{pickSpectra}
\title{Pick peaks in a spectrum}
\usage{
pickSpectra(
  x,
  SNR = 0,
  scales = NULL,
  maxScale = 32,
  noiseWindow = 0.1,
  solventFilter = NULL,
  scaleFactor = 1
)
}
\description{
Pick peaks in a spectrum
}
