% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{pickSpectrum}
\alias{pickSpectrum}
\title{Pick peaks in a spectrum}
\usage{
pickSpectrum(
  x,
  SNR = 0,
  scales = NULL,
  maxScale = 32,
  noiseWindow = 0.1,
  solventFilter = NULL
)
}
\description{
Pick peaks in a spectrum
}
