% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{plot_imp_select}
\alias{plot_imp_select}
\title{feature selection plot}
\usage{
plot_imp_select(feature_select_model)
}
\arguments{
\item{feature_select_model}{feature select model from imp_select function}
}
\description{
feature selection plot
}
