% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA}
\usage{
plot_PCA(data, method = c("facto", "stats"), classifiers)
}
\arguments{
\item{data}{transposed dataframe with m/z as columns}

\item{method}{Which method of calculating the PCA should be used.
"facto" uses the FactoMineR and factoextra packages. "stats" uses prcomp from
the basic stats package and ggpubr.}

\item{classifiers}{sample groups as factor}
}
\value{

}
\description{
visualization using either factoMineR and factorextra package
or stats and ggplot2 .. click next to show the next plot
}
