% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCrossValidation}
\alias{plotCrossValidation}
\title{Plot Accuracy of Cross Validation}
\usage{
plotCrossValidation(exp, modelName = 1)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{modelName}{Either name of number of the model}
}
\description{
Plot Accuracy of Cross Validation
}
