% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(master_df_pos_t, method = c("facto", "stats"))
}
\arguments{
\item{master_df_pos_t}{transposed dataframe}
}
\value{

}
\description{
visualization using either factoMineR and factorextra package
or stats and ggplot2
}
