% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPLSDA}
\alias{plotPLSDA}
\title{Plot PLS-DA}
\usage{
plotPLSDA(data, classifiers, comp, method)
}
\arguments{
\item{data}{transposed dataframe with m/z as columns}

\item{classifiers}{samples groups as factor}

\item{comp}{integer value indicating the component of interest from the object}

\item{method}{method for contribution "median", "mean"}
}
\description{
Plot PLS-DA
}
