% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPlots.R
\name{plotRawFeature}
\alias{plotRawFeature}
\title{Plot feature spectra}
\usage{
plotRawFeature(cells, feature, file)
}
\arguments{
\item{file}{File number of the files that were used.}
}
\description{
Plot feature spectra
}
