% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PLSDA_ROC}
\alias{PLSDA_ROC}
\title{PLSDA ROC plot}
\usage{
PLSDA_ROC(data, classifiers, comp = 1)
}
\arguments{
\item{data}{transposed dataframe with m/z as columns}

\item{classifiers}{sample groups as factor}

\item{comp}{integer value indicating the component of interest from the object (default=1)}
}
\description{
PLSDA ROC plot
}
