% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{ppm_to_dalton}
\alias{ppm_to_dalton}
\title{Dalton error calculation}
\usage{
ppm_to_dalton(mass)
}
\arguments{
\item{mass}{Mass of interest in m/z}

\item{ppm}{Parts Per Million Tolerance}
}
\description{
This function computes the error in Dalton
}
