% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{ppmToDalton}
\alias{ppmToDalton}
\title{Dalton tolerance calculation}
\usage{
ppmToDalton(mass, ppm)
}
\arguments{
\item{mass}{A vector containing mass of interest in m/z}

\item{ppm}{An integer, defining parts per million (ppm) for tolerance (default = 5)}
}
\description{
This function converts the parts per million(ppm)tolerance into Dalton tolerance unit
}
