% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanFilter.R
\name{prepareFile}
\alias{prepareFile}
\title{Format scans in a file}
\usage{
prepareFile(
  file,
  massWindow = c(0, Inf),
  centroid = TRUE,
  polarity = "-",
  combineSpectra = FALSE,
  cl = NULL
)
}
\description{
Format scans in a file
}
