% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{prepareScans}
\alias{prepareScans}
\title{Format scans in files}
\usage{
prepareScans(
  files,
  massWindow = c(0, Inf),
  centroid = TRUE,
  polarity = "-",
  combineSpectra = FALSE
)
}
\description{
Format scans in files
}
