% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{Prob_M}
\alias{Prob_M}
\title{Probability mono-isotopic peak calculation for carbon elements}
\usage{
Prob_M(n)
}
\arguments{
\item{n}{number of atoms}
}
\description{
This function computes the probability of the mono-isotopic peak
}
