% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{Prob_M_1}
\alias{Prob_M_1}
\title{Probability isotopic peak calculation for carbon elements}
\usage{
Prob_M_1(n)
}
\arguments{
\item{n}{number of atoms}
}
\description{
This function computes probability isotope calculation
}
