% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{randomForest_CV}
\alias{randomForest_CV}
\title{feature selection by cross validation}
\usage{
randomForest_CV(training_set, test_set, class1, class2, k, seed, n)
}
\arguments{
\item{training_set}{the training set}

\item{test_set}{the test set}

\item{class1}{sample group 1 name for renaming}

\item{class2}{sample group 2 name for renaming}

\item{k}{number of subsets splitted}

\item{seed}{global seed for reproducible results}

\item{n}{number of imp features to be selected in each model}
}
\description{
This function apply random forest model to the splitted subsets and extract the important variable from each
model and the roc and auc for each model
}
