% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{RandomForestPlots}
\alias{RandomForestPlots}
\title{random forest plots}
\usage{
RandomForestPlots(model, training_set)
}
\arguments{
\item{model}{random forest model}

\item{training_set}{training set that was used for the random forest model with samples column}
}
\description{
random forest plots .. click next to view the next plots
1. OOB plot 2.important variables plot 3.MDS plot
}
