% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rawToMzml}
\alias{rawToMzml}
\title{Convert vendor files to mzML using Proteowizard}
\usage{
rawToMzml(folder, output = getwd(), rt = NULL, options = "--simAsSpectra")
}
\arguments{
\item{folder}{Character path to the folder containing vendor-formatted files,
e.g. wiff from Sciex.}

\item{output}{Folder where to store the resulting mzML files.}

\item{rt}{A 2-sized vector of rt-min and rt-max forming a region to extract
from the vendor-formatted file. Can be used to extract a region of interest.}

\item{options}{Additional options for msConvert. Defaults to
`--simAsSpectra`. Any options given will overwrite this option, unless it is
added to the options.}
}
\description{
Convert vendor files to mzML using Proteowizard
}
