% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{read_msdata}
\alias{read_msdata}
\title{(LEGACY, use read_mzml instead) Read MS data}
\usage{
read_msdata(path = "data")
}
\arguments{
\item{path}{path to the file}
}
\value{
Data.frame with ordered mz values
}
\description{
(LEGACY, use read_mzml instead) Read MS data
}
