% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{read_mzml}
\alias{read_mzml}
\title{Read DIMS data from an mzML file}
\usage{
read_mzml(mzml = NULL)
}
\arguments{
\item{mzml}{Path to a mzml file. If omitted,
the package example will be used}
}
\value{
XCMS object with spectra
}
\description{
Read DIMS data from an mzML file
}
