% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{redvar_removal}
\alias{redvar_removal}
\title{Redundant variables removal}
\usage{
redvar_removal(data, cutoff)
}
\arguments{
\item{data}{the dataframe with mz column and samples columns}

\item{cuttoff}{a certain value of cuttoff for high correlation}
}
\description{
This function removes correlated peaks at a certain cutoff value
}
