% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDefect.R
\name{result_output}
\alias{result_output}
\title{Export results in a csv file}
\usage{
result_output(filtered_df)
}
\arguments{
\item{dataframe}{filtered_MD_df obtained from `MD_filter`}
}
\description{
Export results in a csv file
}
