% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{RF_CV_plots}
\alias{RF_CV_plots}
\title{Plot random forest cross validation ROC}
\usage{
RF_CV_plots(roc_all, auc_all, rocfinal, aucfinal)
}
\arguments{
\item{roc_all}{ROC list from randomForest_CV function output}

\item{auc_all}{AUC list from randomForest_CV function output}

\item{rocfinal}{rocfinal from RF_model function output}

\item{aucfinal}{aucfinal from RF_model function output}
}
\description{
plotting ROC curves for different 5 cross-validation sub models vs ROC curve of the final model
Note : for this plot cross-validation from sumR package needs to be done first
}
