% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{rf_group}
\alias{rf_group}
\title{Random Forest subsets splitting}
\usage{
rf_group(k, data, class1, class2, seed)
}
\arguments{
\item{k}{number of subsets splitted}

\item{data}{the training set}

\item{class1}{sample group 1 name for renaming}

\item{class2}{sample group 2 name for renaming}

\item{seed}{global seed for reproducible results}
}
\description{
This function split the main training set into multiple subsets acoording to K numbers for cross validation
}
