% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{RF_model}
\alias{RF_model}
\title{random forest model}
\usage{
RF_model(training_set, test_set, mtry, ntree, seed)
}
\arguments{
\item{training_set}{the training set}

\item{test_set}{the test set}

\item{mtry}{number of mtry for the model}

\item{ntree}{number of trees for the model}

\item{seed}{global seed for reproducible results}
}
\description{
This function make the final random forest model with the selected features
}
