% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_func.R
\name{roc_binomial}
\alias{roc_binomial}
\title{roc curve for auc}
\usage{
roc_binomial(final_fit, training)
}
\arguments{
\item{final_fit}{the cv.glmnet model with "auc" as type.measure}

\item{training}{training set}
}
\description{
This function makes ROC curve for AUC measure of cv.glmnet
}
