% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ROC_plot}
\alias{ROC_plot}
\title{ROC plot}
\usage{
ROC_plot(roc, auc)
}
\arguments{
\item{roc}{ROC of desired model}

\item{auc}{AUC of desired model}
}
\description{
plotting ROC curves for any choosen model (final (rocfinal and aucfinal) or
from cross validation model(sepcifiy model number roc_all[\link{1}] and auc_all[\link{1}]))
}
