% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveExperiment}
\alias{saveExperiment}
\title{Save a sumR SummarizedExperiment to an RDS file}
\usage{
saveExperiment(exp, path)
}
\arguments{
\item{exp}{SummarizedExperiment obtained after alignment}

\item{path}{Character of the location where to store the Experiment as RDS
file.}
}
\description{
This is a helper function to store the SummarizedExperiment
as an RDS file after validation.
}
\details{
Saving an experiment can be useful for multiple purposes,
including making comparisons and sharing. This function performs a few
validation checks before it creates the directory of \code{path} if it does not
exist yet. After creation, the \link{saveRDS} function is used to store a
compressed file of the SummarizedExperiment to the disk.

Similarly, loading an experiment can be done with the \link{loadExperiment}
function.
}
