% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{setDefaultAssay}
\alias{setDefaultAssay}
\title{Set the default assay in the SummarizedExperiment}
\usage{
setDefaultAssay(exp, default)
}
\arguments{
\item{exp}{SummarizedExperiment object obtained after alignment}

\item{default}{Name of the assay to be set as default assay. Must be
present in \link{assayNames}}
}
\value{
SummarizedExperiment object with the given assay at index 1
}
\description{
This function swaps the current assay at index 1  in \code{exp}
for the assay provided in \code{default}.
}
\details{
By default, most post-processing functions use the assay in \code{exp}
at index 1. While you can set the assayName in these functions, it is often
easier to set the default assay at index 1 instead. This function aids in
doing so by swapping the positions of the assays when calling \link{assayNames}.
}
\examples{
# Read data
data("sumRnegative")

# Impute missing values
sumRnegative <- imputation(sumRnegative, setDefault = FALSE)

# Set the default assay to the Imputed assay
sumRnegative <- setDefaultAssay(sumRnegative, "Imputed")

# Print the assayNames
assayNames(sumRnegative)
}
