% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{shapiro_results}
\alias{shapiro_results}
\title{shapiro test}
\usage{
shapiro_results(dataframe, threshold = 0.05)
}
\arguments{
\item{dataframe}{transposed dataframe with m/z as columns}

\item{threshold}{numerical value for wanted threshold . the default is 0.05}
}
\description{
shapiro test for normality check
}
