% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{shapiroTest}
\alias{shapiroTest}
\title{Shapiro's test for normality across cells}
\usage{
shapiroTest(exp, assay = 1)
}
\arguments{
\item{dataframe}{transposed dataframe with m/z as columns}
}
\description{
shapiro test for normality check
}
