% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{spectraBinning}
\alias{spectraBinning}
\title{Bin spectra with similar masses}
\usage{
spectraBinning(spectraList, ppm = 5)
}
\arguments{
\item{spectraList}{List of spectra to be binned}

\item{ppm}{How much can masses deviate from the mean of the bin?
Defaults to 5 ppm}
}
\value{
DataFrame object from S4Vectors containing aligned spectra
}
\description{
This is an adapted form of Maldiquants binning function. It will
check the given masses and if subsequent masses fall within the ppm
tolerance, it will assign them to the same bin.
}
\details{
Binning is a form of alignment that groups together masses
within a given tolerance. Due to measurements errors in mass spectrometry,
binning is necessary to determine which mass intensities belong to
the same peak.
}
