% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{spectraClustering}
\alias{spectraClustering}
\title{Group spectra using clustering}
\usage{
spectraClustering(peaks, ppm = 5)
}
\arguments{
\item{peaks}{Centroided spectra obtained from \link{extractPeaks}.}

\item{ppm}{Tolerance between masses to be considered part of the same bin
or cluster in parts per million (ppm). Defaults to 5.}
}
\description{
Group spectra using clustering
}
\details{
Binning is a form of alignment that groups together masses
within a given tolerance. Due to measurements errors in mass spectrometry,
binning is necessary to determine which mass intensities belong to
the same peak.
}
