% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPlots.R
\name{spectraShiftPlot}
\alias{spectraShiftPlot}
\title{spectraShiftPlot}
\usage{
spectraShiftPlot(spectraList, file = 1)
}
\arguments{
\item{file}{File number of the files that were used.}
}
\description{
spectraShiftPlot
}
