% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPlots.R
\name{spectrumPlot}
\alias{spectrumPlot}
\title{Plot spectrum peaks}
\usage{
spectrumPlot(peaks, file = 1, scan = 1)
}
\arguments{
\item{peaks}{List of datafres with peak picked data}

\item{file}{File number of the files that were used.}
}
\description{
Plot spectrum peaks
}
