% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{subset_matrix_filter}
\alias{subset_matrix_filter}
\title{Filtering for intensity ratio}
\usage{
subset_matrix(df, mz , Intensity_ratio_vec, by_step)
}
\arguments{
\item{df}{Data frame}

\item{mz}{Mass}

\item{Intensity_ratio_vec}{Intensity ratio vector for carbon}

\item{by_step}{"by" in sequence range}
}
\description{
Filtering for intensity ratio within each range acccording to Intensity ratio vector
}
