% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_function.R
\name{sum_numbers}
\alias{sum_numbers}
\title{sums up a vector of numeric data
Takes in a vector of numeric data and sums them up}
\usage{
sum_numbers(input_data)
}
\arguments{
\item{input_data}{vector of numeric data}
}
\value{
sum of numeric data
}
\description{
sums up a vector of numeric data
Takes in a vector of numeric data and sums them up
}
