% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumRnegative.R
\docType{data}
\name{sumRnegative}
\alias{sumRnegative}
\title{Aligned peaks of single cell DI-MS in negative polarity}
\format{
An object of class \code{SummarizedExperiment} with 841 rows and 8 columns.
}
\usage{
sumRnegative
}
\description{
sumR Package data
}
\keyword{datasets}
