% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{testCheck}
\alias{testCheck}
\title{Check if the test picked can be executed}
\usage{
testCheck(exp, classifiers, minGroups = 0, maxGroups = Inf)
}
\description{
Check if the test picked can be executed
}
