% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{valid_list_as_df}
\alias{valid_list_as_df}
\title{Creating dataframe from isotope_valid_list}
\usage{
valid_list_as_df(isotope_valid_list)
}
\arguments{
\item{isotope_valid_list}{list of annotate information for isotopes}
}
\description{
Creating dataframe from isotope_valid_list
}
