% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateExperiment}
\alias{validateExperiment}
\title{Validate a sumR Experiment on rows, columns and assays}
\usage{
validateExperiment(exp, checkColData = T)
}
\arguments{
\item{exp}{SummarizedExperiment obtained after alignment}
}
\description{
Validate a sumR Experiment on rows, columns and assays
}
