% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validatePeaks}
\alias{validatePeaks}
\title{Validate if a list of centroided data.frames is valid to
use with alignment}
\usage{
validatePeaks(peakList)
}
\arguments{
\item{peakList}{list of data.frames, each with scan, mz, i, and
rt columns.}
}
\description{
This function checks if the centroiding went correctly and
the list of data.frames contain the right columns to be used for
alignment. See details about the validation procedure.
}
\details{
This function checks the list of data.frames on the following
points:
\itemize{
\item Do the names of the list equal the number of entries in the list
\item does each entry is a data.frames with columns \code{scan}, \code{rt}, \code{mz} and \code{i}.
\item Does each entry have the attributes \code{polarity}, \code{massWindow},
\code{combineSpectra} and \code{Files}.
}
}
