% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateSpectra}
\alias{validateSpectra}
\title{Validate the spectra DataFrame}
\usage{
validateSpectra(spectraDf)
}
\arguments{
\item{spectraDf}{DataFrame object obtained from \code{spectraAlignment}}
}
\value{
Either TRUE or FALSE representing the validation of the object.
}
\description{
This function validates if the DataFrame from \code{spectraAlignment}
is suitable for further processing.
}
\details{
\code{spectraAlignment} returns a DataFrame from S4Vectors containing the
following columns:
\itemize{
\item \strong{sample}: Name of the sample the peak originated from
\item \strong{mz}: Median mz of the found peak
\item \strong{mzmin}: Lowest mz of the peak
\item \strong{mzmax}: Highest mz of the peak
\item \strong{npeaks}: Number of scans the peak consists of
\item \strong{i}: (Integrated) Intensity value of the peak
\item \strong{fpeaks}: Fraction of scans the peak is found in. Calculated by dividing
\emph{npeaks} by the total number of scans in the mz region.
\item \strong{ppm}: Parts per million error of the peak. Calculated by
\deqn{ \frac{mzmax - mzmin}{mz} \cdot 1e^6 }.
\item \strong{peakIdx}: List of indexes where the scans are found.
\item \strong{rt}: Median rt of the peak
}
}
