% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{validListAsDf}
\alias{validListAsDf}
\title{Converting isotope_valid_list}
\usage{
validListAsDf(isotope_valid_list)
}
\arguments{
\item{isotope_valid_list}{A list, of valid annotate information concerning the compounds}
}
\description{
Converting isotope_valid_list to a data frame
}
