% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{varImportance}
\alias{varImportance}
\title{Assess the variable importance of models}
\usage{
varImportance(exp, modelName = 1)
}
\arguments{
\item{exp}{SummarizedExperiment with model(s) generated by \link{generateModel}.}

\item{modelName}{Name or index of the model. Defaults to \code{1}, meaning the
first model that was generated.}
}
\value{
A dataframe with two columns: \code{Compound} and \code{Overall}, which
indicates the overall probability of a compound for a model
}
\description{
This function retrieves that variable importances of peaks
that were used in a model. For each peak, the probability is shown in the
\code{Overall} column.
}
\details{
Variable importance is a metric that determines how important the
variable (compound) was for determining the difference between the given
phenotypes. A higher probability indicates a larger difference and a
stronger association with a phenotype.
}
\examples{
# Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Generate a (Random Forest) model
sumRnegative <- generateModel(sumRnegative, "rf", ratio = 0.5)

# Retrieve model infomation
model(sumRnegative)

# Retrieve Variable importances
varImportance(sumRnegative, "rf")
}
