% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot}
\usage{
volcanoPlot(exp, test, title = "")
}
\arguments{
\item{title}{title of the plot}

\item{data}{dataframe result from statistical tests within the sumR package with significant column (true or false)}

\item{xvalues}{log2 folchange between two groups column}

\item{yvalues}{nominal p value column}
}
\value{
volcano plot
}
\description{
Volcano plot
}
