% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_sum.R
\name{weight_sum}
\alias{weight_sum}
\title{Weighted Sum of Numeric vector}
\usage{
weight_sum(x, weights = rep(1, length(x)), ...)
}
\arguments{
\item{x}{numeric vector}

\item{weights}{numeric vector of weight with length of x}

\item{...}{additional imput of sum}
}
\value{
wighted sum
}
\description{
Weighted Sum of Numeric vector
}
\examples{
weight_sum(1:10)
}
