% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{Welch_ttest}
\alias{Welch_ttest}
\title{welch t test}
\usage{
Welch_ttest(
  dataframe,
  dataframe2,
  samples,
  threshold = 0.1,
  corr_method = "fdr"
)
}
\arguments{
\item{dataframe}{transposed dataframe with m/z as columns (transformed dataframe)}

\item{dataframe2}{transposed dataframe with m/z as columns (raw imputed dataframe)}

\item{threshold}{numerical value for wanted threshold . the default is 0.1}

\item{corr_method}{character string for correction method . default is "fdr"}

\item{classifiers}{a factor vector with the classes of the samples}
}
\description{
welch t test between two groups of samples (assuming unequal variance)
}
