% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{welchAnova}
\alias{welchAnova}
\title{welch anova test}
\usage{
welchAnova(
  exp,
  assay = 1,
  classifiers = metadata(exp)$phenotype,
  method = "fdr"
)
}
\description{
welch anova test between more than two groups of samples
(assuming unequal variance)
}
