% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{welchTest}
\alias{welchTest}
\title{welch t test}
\usage{
welchTest(
  exp,
  assay = 1,
  classifiers = metadata(exp)$phenotype,
  method = "fdr"
)
}
\description{
welch t test between two groups of samples (assuming unequal
variance)
}
