% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{wilcox_results}
\alias{wilcox_results}
\title{wilcox test}
\usage{
wilcox_results(
  dataframe,
  dataframe2,
  samples,
  threshold = 0.1,
  corr_method = "fdr",
  paired = F
)
}
\arguments{
\item{dataframe}{transposed dataframe with m/z as columns (transformed dataframe)}

\item{dataframe2}{transposed dataframe with m/z as columns (raw imputed dataframe)}

\item{samples}{a factor vector with the classes of the samples}

\item{threshold}{numerical value for wanted threshold . the default is 0.1}

\item{corr_method}{character string for correction method . default is "fdr"}

\item{paired}{logical vector for paired test .. default paired = FALSE}
}
\description{
wilcox test between two groups of samples
}
