% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_sum.R
\docType{data}
\name{x}
\alias{x}
\title{Mein Datensatz}
\format{
Ein Integer Vektor der Laenge 50
}
\usage{
x
}
\description{
Ein Datensatz bestehend aus 50 Zahlen gezogen aus den
Zahlen 1 bis 100 mit Zuruecklegen
}
\keyword{datasets}
