% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addCellData}
\alias{addCellData}
\title{Add sample data to the experiment}
\usage{
addCellData(exp, data = NULL)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{data}{Dataframe with samplenames as rows}
}
\description{
Add sample data to the experiment
}
