% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addSampleBlankRatio}
\alias{addSampleBlankRatio}
\title{Add the Sample/Blank ratio for each compound}
\usage{
addSampleBlankRatio(exp, assay = 1, typeColumn = "Type")
}
\arguments{
\item{exp}{SummarizedExperiment Object}

\item{typeColumn}{Name of the column in \code{colData(exp)} that represents
the sample type and contains at least the values \code{SAMPLE} and \code{BLANK}.}
}
\value{
A SummarizedExperiment object with results stored in the \link{rowData}
as \code{SampleBlankRatio}.
}
\description{
This function calculates the ratio between the median sample
area and the median blank area. The result is stored in the \link{rowData} slot.
}
\details{
The sample/blank ratio is a metric that aids in determining the
threshold for the blank filter. In cases with a lot of imputed values,
this ratio might be (very) high. In such cases, it might be suitable to
use a log2 or log10 assay to calculate the log fold change instead.
}
\examples{
# Read data
data("sumRnegative")

# Read the cellData
cellData <- read.csv(system.file("cellData.csv", package = "sumR"),
                     sep = ",", row.names = 1)

# Add the cellData
sumRnegative <- addCellData(sumRnegative, cellData)

# Calculate the Sample / Blank ratio
sumRnegative <- addSampleBlankRatio(sumRnegative)

# Show the ratio of the first 5 compounds
rowData(sumRnegative[1:5, ])
}
