% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{align_check}
\alias{align_check}
\title{alignment quality control}
\usage{
align_check(aligned_peaks)
}
\arguments{
\item{aligned_peaks}{dataframe of aligned peaks
obtained from \code{iteration}}
}
\description{
uses the ppm_calc function to check
ppm error of aligned peaks
}
