% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{aovTukeyTest}
\alias{aovTukeyTest}
\title{Anova with Tukey HSD post-hoc test for > 2 groups}
\usage{
aovTukeyTest(exp, assay = 1, classifiers = metadata(exp)$phenotype)
}
\arguments{
\item{exp}{SummarizedExperiment with a (auto)scaled assay.}

\item{assay}{Name or index of the assay with (auto)scaled values. Defaults
to the first assay (index 1).}

\item{classifiers}{Column in the \link{colData} that describes the phenotype or
predictor class used for statistical testing and modelling. Defaults to the
phenotype set using \link{setPhenotype}.}
}
\value{
SummarizedExperiment with \code{aovTukeyTest} column in the \link{rowData}
slot.
}
\description{
This function excutes an anova test and consecutively the
Tukey HSD post-hoc test for determining which compounds are significantly
different across > 2 phenotypes.
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Run Anova & Tukey Test
sumRnegative <- aovTukeyTest(sumRnegative)

# Show results of first 5 compounds
rowData(sumRnegative[1:5, ])$aovTukeyTest
}
