% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{autoTest}
\alias{autoTest}
\title{Detect and execute an appropriate statistical test}
\usage{
autoTest(exp, assay = 1, classifiers = metadata(exp)$phenotype)
}
\description{
Detect and execute an appropriate statistical test
}
