% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Binning function}
\usage{
binning(mass, tolerance, times, sample_num)
}
\arguments{
\item{mass}{}

\item{tolerance}{}

\item{times}{}

\item{sample_num}{}
}
\value{

}
\description{
Binning function
}
