% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binPeaks}
\alias{binPeaks}
\title{Binning of the peaks}
\usage{
binPeaks(df_list, tolerance = 5e-06)
}
\arguments{
\item{df_list}{list of dataframes obtained from the input files}

\item{tolerance}{obtained from the user input or use of default value 5e-6}
}
\description{
binPeaks function! This needs the two functions above
}
