% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking2.R
\name{binSpectra}
\alias{binSpectra}
\title{Bin spectra with similar mass ranges}
\usage{
binSpectra(peakList, fraction = 0, npeaks = 0, meanSNR = 0, tolerance = 0.002)
}
\description{
Bin spectra with similar mass ranges
}
