% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{cellAlignment}
\alias{cellAlignment}
\title{Group cells into SE}
\usage{
cellAlignment(
  spectraDf,
  method = "binning",
  ppm = 5,
  nCells = 0,
  cellData = NULL,
  phenotype = NULL
)
}
\description{
Group cells into SE
}
