% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{cellBinning}
\alias{cellBinning}
\title{Bin cells}
\usage{
cellBinning(spectra, cellData = NULL, phenotype = NULL, ppm = 5)
}
\arguments{
\item{spectraList}{List of spectra}
}
\description{
Bin cells
}
