% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{combineExperiments}
\alias{combineExperiments}
\title{Combine SummarizedExperiments by row}
\usage{
combineExperiments(..., direction = "row", mode = "intersect")
}
\arguments{
\item{...}{Number of SummarizedExperiments}
}
\description{
Combine SummarizedExperiments by row
}
