% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineExperiments}
\alias{combineExperiments}
\title{Combine SummarizedExperiments by row}
\usage{
combineExperiments(exps)
}
\arguments{
\item{exps}{A list of SummarizedExperiments to be combined}
}
\description{
Combine SummarizedExperiments by row
}
