% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{data_transform}
\alias{data_transform}
\title{data transformation}
\usage{
data_transform(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
This function apply data transformation using log2 and pareto scale
}
