% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{detectTest}
\alias{detectTest}
\title{Detect an appropriate statistical test}
\usage{
detectTest(
  exp,
  assay = 1,
  classifiers = metadata(exp)$phenotype,
  threshold = 2/3
)
}
\description{
Detect an appropriate statistical test
}
