% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{doBinning}
\alias{doBinning}
\title{Bin masses of spectra}
\usage{
doBinning(spectra, split = "scan", ppm = 5)
}
\arguments{
\item{spectra}{a data.frame or list of data.frames containing centroided
spectra.}

\item{split}{In case of a data.frame in \code{spectra}, which column should
be used to indicate different scans on which the alignment should occur?
Defaults to "scan"}

\item{ppm}{How much can masses deviate from the mean of the bin?
Defaults to 5 ppm}
}
\value{
A data.frame with the binned masses, their intensities and
original peak identifiers for traceback.
}
\description{
This is an adapted form of Maldiquants binning function. It will
check the given masses and if subsequent masses fall within the ppm
tolerance, it will assign them to the same bin.
}
\details{
Binning is a form of alignment that groups together masses
within a given tolerance. Due to measurements errors in mass spectrometry,
binning is necessary to determine which mass intensities belong to
the same peak.
}
