% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{Dunn_test}
\alias{Dunn_test}
\title{Dunn test}
\usage{
Dunn_test(dataframe, samples, threshold, comb, method = "bh")
}
\arguments{
\item{dataframe}{transposed dataframe with m/z as columns (results of significant kruskal test)}

\item{threshold}{numerical value for wanted threshold}

\item{comb}{a numerical value for number of group combinations}

\item{method}{a character string for correction method .. default is "bh"}

\item{classifiers}{a factor vector with the classes of the samples}
}
\description{
Dunn post hoc test after kruskal-wallis test
}
