% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{dunnTest}
\alias{dunnTest}
\title{Dunn test}
\usage{
dunnTest(
  exp,
  classifiers,
  assay = 1,
  threshold = 0.05,
  comb = 6,
  method = "bh"
)
}
\arguments{
\item{threshold}{numerical value for wanted threshold default is 0.05}

\item{comb}{a numerical value for number of group combinations (4 groups = 6 combination)}

\item{method}{a character string for correction method .. default is "bh"}

\item{dataframe}{transposed dataframe with m/z as columns (results of significant kruskal test)}

\item{samples}{a factor vector with the classes of the samples}
}
\description{
Dunn post hoc test after kruskal-wallis test
}
