% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{dunnTest}
\alias{dunnTest}
\title{Dunn test}
\usage{
dunnTest(exp, assay = 1, classifiers = metadata(exp)$phenotype, method = "bh")
}
\arguments{
\item{exp}{SummarizedExperiment with a (auto)scaled assay.}

\item{assay}{Name or index of the assay with (auto)scaled values. Defaults
to the first assay (index 1).}

\item{classifiers}{Column in the \link{colData} that describes the phenotype or
predictor class used for statistical testing and modelling. Defaults to the
phenotype set using \link{setPhenotype}.}

\item{Correction}{method to address the multiple testing problem.
Defaults to \code{"bh"} (Benjamini-Hochberg).}
}
\value{
SummarizedExperiment with \code{dunnTest} column in the \link{rowData} slot.
}
\description{
Dunn post hoc test after kruskal-wallis test
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Run Dunn's Test
sumRnegative <- dunnTest(sumRnegative)

# Show results of first 5 compounds
rowData(sumRnegative[1:5, ])$dunnTest
}
\seealso{
\link{dunn.test} for correction methods available.
}
