% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{feature_processing}
\alias{feature_processing}
\title{Process features obtained from XCMS}
\usage{
feature_processing(data)
}
\arguments{
\item{data}{XCMS object with grouped peaks}
}
\description{
Process features obtained from XCMS
}
