% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{foldChange}
\alias{foldChange}
\title{foldchange for two groups test}
\usage{
foldChange(exp, classifiers = metadata(exp)$phenotype, assay = 1)
}
\arguments{
\item{dataframe2}{transposed dataframe with m/z as columns (raw imputed dataframe)}

\item{samples}{a factor vector with the classes of the samples}
}
\description{
foldchange for two groups test
}
