% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{foldChange}
\alias{foldChange}
\title{foldchange for two groups test}
\usage{
foldChange(exp, assay = 1, classifiers = metadata(exp)$phenotype)
}
\arguments{
\item{classifiers}{a factor vector with the classes of the samples}
}
\description{
foldchange for two groups test
}
