% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{generateModel}
\alias{generateModel}
\title{Model using RandomForest model}
\usage{
generateModel(
  exp,
  modelName,
  classifiers = metadata(exp)$phenotype,
  assay = 1,
  cv = 5,
  ratio = 0.632,
  seed = NULL,
  ...
)
}
\description{
Model using RandomForest model
}
