% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{generateModel}
\alias{generateModel}
\title{Model phenotypes using the caret library}
\usage{
generateModel(
  exp,
  modelName,
  classifiers = metadata(exp)$phenotype,
  assay = 1,
  folds = 5,
  ratio = 0.632,
  seed = NULL,
  ...
)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{classifiers}{Column name in the colData that represents the classifiers}

\item{assay}{Which assay should be used to create a model}

\item{ratio}{Train-test split ratio}

\item{...}{Any other parameters for preProcessing}

\item{cv}{Number of folds to be used for cross validation}
}
\value{
SummarizedExperiment with updated model field in the metadata slot.
}
\description{
This function trains and tests a given model from the caret
library to predict phenotypes. This can be used as a multivariate approach
to determine important differentiating peaks between phenotypes.
}
\details{
Modelling is a multi-variate approach of determining important
peaks to predict the phenotype(s). Models are created using the \code{caret}
package. First, a train-test split is made using the given ratio. Next,
train control is done using a cross-validation approach. Here, the data
is scaled and centered before training starts. The model used for training
can be picked from the caret library of classification models. The results
are stored in the model field of the SummarizedExperiment metadata slot. It
can be accessed by the \link{model} function.
}
\examples{
# Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Generate a (Random Forest) model
sumRnegative <- generateModel(sumRnegative, "rf", ratio = 0.5)

# Retrieve model infomation
model(sumRnegative)

# Retrieve Variable importances
varImportance(sumRnegative, "rf")
}
\seealso{
\href{https://topepo.github.io/caret/}{Caret manual}
}
