% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Impute missing values}
\usage{
imputation(
  exp,
  method = "noise",
  useAssay = "Area",
  saveAssay = "Imputed",
  setDefault = TRUE,
  ...
)
}
\description{
Impute missing values
}
