% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_tagging.R
\name{isotope_tagging}
\alias{isotope_tagging}
\title{Isotope tagging}
\usage{
Isotope_tagging(df, iso_diff , ppm , by_step )
}
\arguments{
\item{df}{Dataframe must contain columns c("mz", "intensity") with "mz" as first column}

\item{iso_diff_da}{Mass difference between mol ion and isotope}

\item{ppm}{Parts Per Million Tolerance}

\item{by_step}{"by" in sequence mz range}
}
\description{
Tags isotopes based on intensity ratio and difference in mass
}
