% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{iteration}
\alias{iteration}
\title{iterations of the alignment function over the data}
\usage{
iteration(df_list, max_align = 8, bin_plot = F)
}
\arguments{
\item{df_list}{list of dataframes obtained from `binPeaks`}

\item{max_align}{value obtained from user input or use of default value 8}

\item{bin_plot}{logical value obtained from user input per default set to FALSE
decides if the line plot showing the reduction of the bins per iteration is shown}
}
\description{
the code iterates over the data as long as
the number of bins changes, when the number of bins
doesn't change anymore iterations stop, maximal number
of iterations is set to 15
Also, plotting the decrease of the bins
}
