% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{leveneTest}
\alias{leveneTest}
\title{Levene test for unequal variances across phenotypes}
\usage{
leveneTest(
  exp,
  assay = 1,
  classifiers = metadata(exp)$phenotype,
  method = "fdr"
)
}
\arguments{
\item{exp}{SummarizedExperiment with a (auto)scaled assay.}

\item{assay}{Name or index of the assay with (auto)scaled values. Defaults
to the first assay (index 1).}

\item{classifiers}{Column in the \link{colData} that describes the phenotype or
predictor class used for statistical testing and modelling. Defaults to the
phenotype set using \link{setPhenotype}.}

\item{method}{Correction method to address the multiple testing problem.
Defaults to \code{"fdr"} (False Discovery Rate).}
}
\value{
SummarizedExperiment with \code{leveneTest} column in the \link{rowData} slot.
}
\description{
This function tests the variance across phenotypes for each
peak and adds the corresponding p-value to the column \code{leveneTest} to the
\link{rowData} slot.
}
\details{
The levene test is a test for calculating unequal variances across
groups. This can be an indicator for peaks that have different values in the
provided assay.
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Run Levene's Test
sumRnegative <- leveneTest(sumRnegative)

# Show results of first 5 compounds
rowData(sumRnegative[1:5, ])$leveneTest
}
