% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{leveneTest}
\alias{leveneTest}
\title{Levene test for unequal variances across phenotypes}
\usage{
leveneTest(exp, assay = 1, classifiers = metadata(exp)$phenotype)
}
\arguments{
\item{exp}{transposed dataframe with m/z as columns}

\item{classifiers}{a factor vector with the classes of the samples}
}
\description{
levene test for variance check
}
