% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{leveneTest}
\alias{leveneTest}
\title{levene test}
\usage{
leveneTest(exp, classifiers, assay = 1, threshold = 0.05)
}
\arguments{
\item{classifiers}{a factor vector with the classes of the samples}

\item{threshold}{numerical value for wanted threshold . the default is 0.05}

\item{dataframe}{transposed dataframe with m/z as columns}
}
\description{
levene test for variance check
}
