% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadExperiment}
\alias{loadExperiment}
\title{Load a sumR Experiment from a file}
\usage{
loadExperiment(path)
}
\arguments{
\item{path}{Character path to the location where the RDS file is stored.}
}
\description{
SummarizedExperiments can be saved during any step after
alignment using the `saveExperiment` function. This function reads the
resulting RDS file that saveExperiment produces. This allows for saving
work and continuing it later.
}
