% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDefect.R
\name{make_filter_list}
\alias{make_filter_list}
\title{Defining Mass Defect filter}
\usage{
make_filter_list(MD_df)
}
\arguments{
\item{dataframe}{MD_df obtained from the HMDB using `mass_defect_function`}
}
\description{
the theoretical maximum Mass Defect of human metabolites is used to calculate this cut off
}
