% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDefect.R
\name{mass_defect_calculation}
\alias{mass_defect_calculation}
\title{Mass defect (MD) calculation}
\usage{
mass_defect_calculation(dataframe)
}
\arguments{
\item{dataframe}{data frame containing experimental data}
}
\description{
here a simplified MD, the mass' decimal number
}
