% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDefect.R
\name{massDefectFilter}
\alias{massDefectFilter}
\title{Mass defect filter pipeline}
\usage{
massDefectFilter(
  exp,
  incl_list = FALSE,
  incl_list_path = system.file("extdata/hmdb_inclusions_list_pos_McMillan.txt", package
    = "sumR"),
  mass_accuracy = 0.01
)
}
\arguments{
\item{dataframe}{containing the data, intensities and mz}

\item{mz_MD_plot}{logical variable deciding if plot is
created per default set to TRUE}
}
\description{
uses the mass defect functions to filter
the data with one command
}
