% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{MD_filter}
\alias{MD_filter}
\title{MD filter}
\usage{
MD_filter(dataframe, mz_col, a = 0.00112, b = 0.01953)
}
\arguments{
\item{dataframe}{is the dataframe on which KM calculations need to be
conducted}

\item{mz_col}{is the mass to charge column within the specified dataframe}

\item{a}{is the coefficient of the linear equation used for data filtering}

\item{b}{is the addition value of the linear equation used for data
filtering, both a and b can be calculated using the linear_equation()
function included in this package.}
}
\description{
filters data based on Mass Defect
}
\examples{
MD_filter(master_df_pos, master_df_pos$mz)

}
