% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{model<-}
\alias{model<-}
\title{Set or replace a model in the given experiment}
\usage{
model(exp, modelName) <- value
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{modelName}{Name of the model to adjust}

\item{value}{Value to set in the model}
}
\description{
Set or replace a model in the given experiment
}
