% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{model}
\alias{model}
\title{Get model in Experiment}
\usage{
model(exp, modelName = 1)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{modelName}{Name of the model to retrieve}
}
\description{
After modelling, the results are stored in the \link{metadata} slot
as a list with name \code{model}. However, multiple models can be made and stored
this way. This function returns the model information of the given model,
either by index or name of the model used.
}
