% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PCA_ellipse}
\alias{PCA_ellipse}
\title{PCA ellipse plot}
\usage{
PCA_ellipse(exp, classifiers, assay = 1)
}
\arguments{
\item{classifiers}{sample groups as factor}

\item{data}{transposed dataframe with m/z as columns}
}
\description{
PCA ellipse plot
}
