% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PCA_ellipse_stats}
\alias{PCA_ellipse_stats}
\title{PCA ellipse plot from stats}
\usage{
PCA_ellipse_stats(data, classifiers)
}
\arguments{
\item{data}{transposed dataframe with m/z as columns}

\item{classifiers}{sample groups as factor}
}
\description{
PCA ellipse plot from stats
}
