% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{Perform peak picking}
\usage{
peakPicking(
  fileList,
  maxScale = 32,
  noiseWindow = 0.1,
  cores = 1,
  solventFilter = NULL,
  scales = NULL
)
}
\description{
Perform peak picking
}
