% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{Perform peak picking}
\usage{
peakPicking(
  files,
  massDefect = 0.8,
  polarity = NULL,
  massWindow = Inf,
  cores = detectCores(logical = F),
  SNR = 0
)
}
\description{
Perform peak picking
}
