% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{Perform peak picking}
\usage{
peakPicking(
  files,
  polarity = NULL,
  combineSpectra = FALSE,
  massWindow = c(0, Inf),
  noiseWindow = 0.1,
  cores = detectCores(logical = F)
)
}
\description{
Perform peak picking
}
