% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot Principle Component Analysis}
\usage{
plotPCA(se, assay = 1, group = metadata(se)$phenotype)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay}{Assay to use}

\item{colors}{optional vector of colors}
}
\description{

}
\details{

}
