% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PLSDA_ind}
\alias{PLSDA_ind}
\title{PLSDA individual plot}
\usage{
PLSDA_ind(exp, classifiers, assay = 1)
}
\arguments{
\item{classifiers}{sample groups as factor}

\item{data}{transposed dataframe with m/z as columns}
}
\description{
PLSDA individual plot
}
