% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/binning.R
\name{ppm_calc}
\alias{ppm_calc}
\title{ppm calculation}
\usage{
ppm_calc(mass1, mass2)

ppm_calc(mass1, mass2)
}
\arguments{
\item{mass1}{obtained from the input files}

\item{mass2}{obtained from the input files}
}
\description{
calculates the parts per million error between two different
masses

ppm_calc calculated the parts per million error between two different masses
}
\examples{
ppm_calc(mass1, mass2)
}
