% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ppm_err_plot}
\alias{ppm_err_plot}
\title{Boxplot of the ppm errors}
\usage{
ppm_err_plot(ppm_err_fn)
}
\arguments{
\item{ppm_err_fn}{dataframe obtained from \code{align_check}}
}
\description{
Boxplot of the ppm errors
}
