% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF_functions.R
\name{redvar_removal}
\alias{redvar_removal}
\title{Redundant variables removal}
\usage{
redvar_removal(data, cutoff = 0.75)
}
\arguments{
\item{data}{the dataframe with m/z as columns and a sample column named samples}

\item{cuttoff}{a certain value of cuttoff for high correlation (default=0.75)}
}
\description{
This function removes correlated peaks at a certain cutoff value
}
