% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{RF_CV_plots}
\alias{RF_CV_plots}
\title{Plot random forest cross validation ROC}
\usage{
RF_CV_plots(model, test_set)
}
\arguments{
\item{model}{the final model}

\item{test_set}{the final test set with samples column}
}
\description{
plotting ROC curves for different 5 cross-validation sub models vs ROC curve of the final model
Note : for this plot cross-validation from sumR package needs to be done first
}
