% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveExperiment}
\alias{saveExperiment}
\title{Save a sumR Experiment to a file}
\usage{
saveExperiment(exp, path)
}
\arguments{
\item{exp}{SummarizedExperiment obtained after alignment}

\item{path}{Character of the location where to store the Experiment as RDS
file.}
}
\description{
Save a sumR Experiment to a file
}
