% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{setPhenotype}
\alias{setPhenotype}
\title{Set the phenotype column of a sumR Experiment}
\usage{
setPhenotype(exp, phenotype)
}
\arguments{
\item{exp}{SummarizedExperiment obtained after alignment}

\item{phenotype}{Character vector specifying the column in colData to
be used as phenotype}
}
\description{
Set the phenotype column of a sumR Experiment
}
