% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{shapiroTest}
\alias{shapiroTest}
\title{shapiro test}
\usage{
shapiroTest(exp, assay = 1, threshold = 0.05)
}
\arguments{
\item{threshold}{numerical value for wanted threshold . the default is 0.05}

\item{dataframe}{transposed dataframe with m/z as columns}
}
\description{
shapiro test for normality check
}
