% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{shapiroTest}
\alias{shapiroTest}
\title{Shapiro's test for normality across cells}
\usage{
shapiroTest(exp, assay = 1, method = "fdr")
}
\arguments{
\item{exp}{SummarizedExperiment with a (auto)scaled assay.}

\item{assay}{Name or index of the assay with (auto)scaled values. Defaults
to the first assay (index 1).}

\item{method}{Correction method to address the multiple testing problem.
Defaults to \code{"fdr"} (False Discovery Rate).}
}
\value{
SummarizedExperiment with updated \link{rowData} slot. The results will
be stored in the \code{shapiroWilk} column of the rowData of \code{exp}.
}
\description{
Shapiro's test for normality is a test that indicates if the
supplied distribution is significantly different from a normal distribution.
This test stored the results of the adjusted p-value in the \link{rowData} slot
of the \code{exp}.
}
\details{
A significant adjusted p-value indicates that the distribution is
very likely not normal, and therefore non-parametric tests should be used to
get an accurate indication of significance, e.g. \link{wilcoxTest} for 2 groups
and \link{dunnTest} for > 2 groups.

It is recommended to impute an assay using e.g. \link{imputation} and scale it
using \link{autoScale} to get a fair estimation of normality.
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Run Shapiro's test
sumRnegative <- shapiroTest(sumRnegative)

# Show results of first 5 compounds
rowData(sumRnegative[1:5, ])$shapiroTest
}
