% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{spectraBinning}
\alias{spectraBinning}
\title{Bin spectra with similar mass ranges}
\usage{
spectraBinning(
  peakList,
  fraction = 0,
  npeaks = 0,
  method = "sum",
  meanSNR = 0,
  tolerance = 0.002
)
}
\description{
Bin spectra with similar mass ranges
}
