% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{spectraBinning}
\alias{spectraBinning}
\title{Bin spectra with similar mass ranges}
\usage{
spectraBinning(spectraList, ppm = 5)
}
\arguments{
\item{spectraList}{List of spectra to be binned}

\item{ppm}{How much can masses deviate from the mean of the bin?
Defaults to 5 ppm}
}
\value{
DataFrame object from S4Vectors containing aligned spectra
}
\description{
Bin spectra with similar mass ranges
}
\details{
Binning is a form of alignment that groups together masses
within a given tolerance. Due to measurements errors in mass spectrometry,
binning is necessary to determine which mass intensities belong to
the same peak.
}
