% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGrouping.R
\name{spectraClustering}
\alias{spectraClustering}
\title{Group spectra using clustering}
\usage{
spectraClustering(spectraList, ppm = 5)
}
\arguments{
\item{spectraList}{List of spectra to be binned}

\item{ppm}{How much can masses deviate from the mean of the cluster?
Defaults to 5 ppm}
}
\description{
Group spectra using clustering
}
\details{
Binning is a form of alignment that groups together masses
within a given tolerance. Due to measurements errors in mass spectrometry,
binning is necessary to determine which mass intensities belong to
the same peak.
}
