% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumR-package.R
\docType{package}
\name{sumR-package}
\alias{sumR}
\alias{sumR-package}
\title{sumR Package}
\description{
sumR Package
}
\author{
\strong{Maintainer}: Pascal Maas \email{p.maas@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0001-9379-6041}{ORCID})

Authors:
\itemize{
  \item Alida Kindt \email{a.s.d.kindt@lacdr.leidenuniv.nl}
  \item Ahmed Ali \email{ali@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0003-2157-4399}{ORCID})
  \item Klara Blaschitz \email{klara.blaschitz@outlook.com} (\href{https://orcid.org/0000-0001-5059-4730}{ORCID})
  \item Eman Muhammad
  \item Merys Valdez \email{m.e.valdez@lacdr.leidenuniv.nl}
  \item David Rasp \email{d.rasp@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0003-0164-2163}{ORCID})
}

}
\keyword{internal}
