% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumR.R, R/sumRC.R
\name{sumR}
\alias{sumR}
\alias{sumRC}
\title{Addition of two numbers.}
\usage{
sumR(x, y)

sumRC(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The addition of \code{x} and \code{y}.
}
\description{
This package contains several functions to compute the addition
of two numbers.
}
\details{
\code{sumRC} is an implementation using the \code{.C} interface.
}
\examples{
sumR(1, 2)

sumRC(1, 2)

}
