% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/sumR.R, R/sumRC.R,
%   R/sumRCall.R
\name{sumCpp}
\alias{sumCpp}
\alias{sumR}
\alias{sumRC}
\alias{sumRCall}
\title{Addition of two numbers.}
\usage{
sumCpp(x, y)

sumR(x, y)

sumRC(x, y)

sumRCall(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The addition of \code{x} and \code{y}.
}
\description{
This package contains several functions to compute the addition
of two numbers.
}
\details{
\code{sumCpp} is an implementation using the Rcpp interface.

\code{sumRC} is an implementation using the \code{.C} interface.

\code{sumRCall} is an implementation using the \code{.Call} interface.
}
\examples{
sumCpp(1, 2)

sumR(1, 2)

sumRC(1, 2)

sumRCall(1, 2)

}
