% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumR.R
\name{sumR}
\alias{sumR}
\title{Addition of two numbers.}
\usage{
sumR(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The addition of \code{x} and \code{y}.
}
\description{
Addition of two numbers.
}
\examples{
sumR(1, 2)

}
