% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{testCheck}
\alias{testCheck}
\title{Check if the test picked can be executed}
\usage{
testCheck(exp, classifiers, minGroups = 0, maxGroups = Inf)
}
\arguments{
\item{exp}{SummarizedExperiment with a (auto)scaled assay.}

\item{classifiers}{Column in the \link{colData} that describes the phenotype or
predictor class used for statistical testing and modelling. Defaults to the
phenotype set using \link{setPhenotype}.}

\item{minGroups}{Minimum number of groups for the test. Defaults to \code{0}}

\item{maxGroups}{Maximum number of groups for the test. Defaults to \code{Inf}}
}
\value{
Boolean value, is the test valid to use?
}
\description{
This function checks if the test is suitable for the data
that is provided and will throw a warning if not or and error when no
classifiers are provided.
}
\details{
This function is called by all tests to see if the test is
appropriate as a check. Used internally by sumR.
}
