% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{varImportance}
\alias{varImportance}
\title{Variable Importance of models}
\usage{
varImportance(exp, modelName = 1)
}
\arguments{
\item{exp}{SummarizedExperiment with model(s)}

\item{modelName}{name of the model}
}
\description{
Variable Importance of models
}
