% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot}
\usage{
volcanoPlot(data, xvalues, yvalues, title)
}
\arguments{
\item{data}{}

\item{xvalues}{}

\item{yvalues}{}

\item{title}{}
}
\value{

}
\description{
Volcano plot
}
