% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{welchAnova}
\alias{welchAnova}
\title{welch anova test}
\usage{
welchAnova(exp, classifiers, assay = 1, threshold = 0.1)
}
\arguments{
\item{threshold}{numerical value for wanted threshold . the default is 0.1}

\item{dataframe}{transposed dataframe with m/z as columns (transformed dataframe)}

\item{dataframe2}{transposed dataframe with m/z as columns (raw imputed dataframe)}

\item{samples}{a factor vector with the classes of the samples}
}
\description{
welch anova test between more than two groups of samples (assuming unequal variance)
}
