% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{WelchAnova}
\alias{WelchAnova}
\title{welch anova test}
\usage{
WelchAnova(dataframe, dataframe2, samples, threshold = 0.1)
}
\arguments{
\item{dataframe}{transposed dataframe with m/z as columns (transformed dataframe)}

\item{dataframe2}{transposed dataframe with m/z as columns (raw imputed dataframe)}

\item{samples}{}

\item{threshold}{numerical value for wanted threshold . the default is 0.1}
}
\description{
welch anova test between more than two groups of samples (assuming unequal variance)
}
