% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{welchTest}
\alias{welchTest}
\title{welch t test}
\usage{
welchTest(
  exp,
  assay = 1,
  classifiers = metadata(exp)$phenotype,
  threshold = 0.1,
  corr_method = "fdr"
)
}
\arguments{
\item{threshold}{numerical value for wanted threshold . the default is 0.1}

\item{corr_method}{character string for correction method . default is "fdr"}
}
\description{
welch t test between two groups of samples (assuming unequal variance)
}
