% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_functions.R
\name{welchTest}
\alias{welchTest}
\title{welch t test}
\usage{
welchTest(exp, classifiers, assay = 1, threshold = 0.1, corr_method = "fdr")
}
\arguments{
\item{threshold}{numerical value for wanted threshold . the default is 0.1}

\item{corr_method}{character string for correction method . default is "fdr"}

\item{dataframe}{transposed dataframe with m/z as columns (transformed dataframe)}

\item{dataframe2}{transposed dataframe with m/z as columns (raw imputed dataframe)}

\item{samples}{a factor vector with the classes of the samples}
}
\description{
welch t test between two groups of samples (assuming unequal variance)
}
