% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{welchTest}
\alias{welchTest}
\title{Welch T-test for 2 groups}
\usage{
welchTest(
  exp,
  assay = 1,
  classifiers = metadata(exp)$phenotype,
  method = "fdr"
)
}
\arguments{
\item{exp}{SummarizedExperiment with a (auto)scaled assay.}

\item{assay}{Name or index of the assay with (auto)scaled values. Defaults
to the first assay (index 1).}

\item{classifiers}{Column in the \link{colData} that describes the phenotype or
predictor class used for statistical testing and modelling. Defaults to the
phenotype set using \link{setPhenotype}.}

\item{method}{Correction method to address the multiple testing problem.
Defaults to \code{"fdr"} (False Discovery Rate).}
}
\value{
SummarizedExperiment with \code{welchTest} column in the \link{rowData} slot.
}
\description{
Welch T-test between two groups of samples (assuming unequal
variance).
}
\details{
The Welch T-test is a two-sample test to see if the two
phenotypes have equal means in their distribution. This test assums unequal
variance across the phenotypes.
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Run Welch' T-Test
sumRnegative <- welchTest(sumRnegative)

# Show results of first 5 compounds
rowData(sumRnegative[1:5, ])$welchTest
}
