% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{wilcoxTest}
\alias{wilcoxTest}
\title{Wilcoxon test for non-parametric distributed data of 2 groups}
\usage{
wilcoxTest(
  exp,
  assay = 1,
  classifiers = metadata(exp)$phenotype,
  method = "fdr"
)
}
\arguments{
\item{exp}{SummarizedExperiment with a (auto)scaled assay.}

\item{assay}{Name or index of the assay with (auto)scaled values. Defaults
to the first assay (index 1).}

\item{classifiers}{Column in the \link{colData} that describes the phenotype or
predictor class used for statistical testing and modelling. Defaults to the
phenotype set using \link{setPhenotype}.}

\item{method}{Correction method to address the multiple testing problem.
Defaults to \code{"fdr"} (False Discovery Rate).}
}
\value{
SummarizedExperiment with \code{wilcoxTest} column in the \link{rowData} slot.
}
\description{
This function will perform the Wilcoxon signed-rank test for
non-parametric data. The groups are determined by the phenotype given and
will produce an error if the number of groups is not equal to 2.
}
\details{
This test is a common test that tests the difference between the
phenotypes proivded. It is considered the equivalent to the \link{welchTest}, but
for non-parametric data, which causes it to be less powerful in general.
}
\examples{
#' # Read example data
data("sumRnegative")

# Set colData and phenotype
df <- read.csv(system.file("cellData.csv", package = "sumR"), row.names = 1)
sumRnegative <- addCellData(sumRnegative, df)
sumRnegative <- setPhenotype(sumRnegative, "Treatment")

# Imputate and Scale data
sumRnegative <- imputation(sumRnegative)
sumRnegative <- autoScale(sumRnegative)

# Run Wilcoxon Test
sumRnegative <- wilcoxTest(sumRnegative)

# Show results of first 5 compounds
rowData(sumRnegative[1:5, ])$wilcoxTest
}
